==> draft/from_FOA_folder/06A_duct_data_analysis.R <==
#
# 06A_duct_data_analysis.R
#
# July 7, 2015: for final AL EUI analysis (based on discussion in a meeting on July 6, 2015)
#               (a) stay with the fixed 1500 samples for each CZ
#               (b) not providing a state EUI cdf which is not possible for the fixed 1500 samples for each CZ
#               (c) providing a weighted histogram through weighted count in the bins
# Note: (1) "2015Jun15": 		the data downloaded date
#       (2) "AL_June29_2015_Final": 	simulation folder
#               
# June 9, 2015: Mark (June 05, 2015 1:54 PM) provided the CZ fraction at AL based on permit (see Y:\FOA\Phase1_Analysis\RCD_Analysis\00_auxi_files\Alabama Monthly Overall v3 (with counties listed) and climate zones.xlsx"
# where 3A is 77.65% and 2A is 22.35%
# hard coded these.
# 
# May 29, 2015: start with AL simulation.
# The number of simulation runs at Al is:
#    2 (CZ)             * 4 (found) * 4 (htgSys) = 32    TRB
# 1500 (random drawing) * 4 (found) * 4 (htgSys) = 24000 PRO
#
# April 30, 2015: to analyze the E+ simulation results for Phase 1
#
# The simululation of phase1 includes the following:
# (1) the survey data for around 200 data items including 18 simulation needed key code item which corresponding to the nine code item considered in the simulation of single family house.
# (2) there are 8 states involved in this FOA phase 1 round each consists of one to three CZs.
# (3) The 63 minimum sample size was based on the overall state level delta EUI standard deviation analysis and therefore shold apply to the state level observations
# (4) random drawing observations on either stae level observations or subsetted CZ of the state level observations need to be re-considered.
# (5) it is planned to draw 1500 houses for each state for the simulation study
# (6) for each of the N = 1500 randomly drawn houses, the nine code items are simulatenously drawn, togther with other parameters fixced at the IECC code level, each draw characterizes one house in the state.
# (7) the N=1500 houses is split to the CZs of the state proportionally to the CZ shares in the state.
# (8) for each M = CZ of the state, a code complance house is specified.
# (9) for each of the N+M houses where N is the number of randomly drawn houses and M is the number of CZs in the state, 16 variants are generated by using one of the 4 heating systems and one of the foundation.
# (10) total number of E+ simulation models is (N + M) X 16 for each state.
#
#
# Data Analysis
# (1) The summary file of E+ is loaded which consists of all (N+M) X 16 models
# (2) the enduse / fuel type need to be combined
# (3) the EUI DIFF (enduse and the whole building) between each each PRO model and its correesponding baseline model are calculated,
# (4) the htg systems and foundation shares in the state are attached (note: they are independent from the CZ in the states).
# (5) weighted averages are calculated on the EUI DIFF 
#
# August 30, 2014
#
# August 25, 2014
# This script is to analyze the simulation data from the Monte Carlo sample (about ~6000 - 7000 models)
# The models runs are conducted on 
# 1) nine measures which are 	i) "r_ceiling", ii) "r_wall", iii) "r_floor", iv) "u_window", v) "shgc_window", vi) "f_inc_hw", vii) "ach50", viii) "r_bsmtwall",and ix) "r_slab"
#    the value range of the nine code items vary from CZ (Mark provided value range estimates, New Expected Range 8-19-2014 v2 Delphi Input.xlsx, Tuesday, August 19, 2014 10:26 AM)
# 2) foundations: 		i) "heatedbsmt", ii) "unheatedbsmt", iii) "crawlspace", and iv) "slab"
# 3) four htgSys: 		i) "gasfurnace", ii) "electricfurnace", iii) "heatpump", and iv) "oilfurnace")
# 4) eight climate cities:	i) "Houston",ii) "Phoenix", iii) "Memphis", iv) "ElPaso", v) "Baltimore", vi) "Albuquerque", vii) "Chicago", and viii) "Burlington")
# 5) eight states:		i) PA, ii) MD, iii) NC, iv) AL, v) GA, vi) AR, vii) TX, viii) KY
#    map state to climate city:
#    i) PA --> 		         | ii) MD --> 		   | iii) NC -->              | iv) AL --> 2A Houston | v) GA --> 2A Houston   | vi) AR --> 	          | vii)TX --> 2A Houston     | vii) KY --> 		          | (3) 
#				 | 			   | 			      | 		      | 	               |  		          |            2B Phoenix     | 				  | (1)
#		  	   	 | 	   		   | 	         3A Memphis   | 	   3A Memphis | 	  3A Memphis   | 	    3A Memphis	  |            3A Memphis     | 				  | (5)
#				 | 			   | 			      | 		      | 		       |  			  |            3B El Paso     | 				  | (1)
#		  4A Baltimore	 | 	      4A Baltimore | 		 4A Baltimore | 		      | 	  4A Baltimore | 	    4A Baltimore  |                           | 		    4A Baltimore  | (6)
#		          	 | 			   | 			      | 		      | 		       | 			  | 	       4B Albuquerque | 				  | (1)
#		  5A Chicago	 | 	      5A Chicago   | 		 5A Chicago   | 		      | 		       | 			  | 			      | 				  | (3)
#		  6A Burlington	 | 			   | 			      | 		      | 		       | 			  | 			      | 				  | (1)
# 6) weights:
#    weights used in the previous 2015 IECC determination (May 17, 2014): 		\\poncho\resstd\determination\IECC2015\quantitative\analysis\documentation\weights\weights.xlsx.
#    weights used in the 2014 May-June for 2015 IECC determination (June 6, 2014): 	\\poncho\resstd\determination\IECC2015\quantitative\analysis\00_weights	   (several csv files generated: revised_weights_single_CZ_foundation.csv, revised_weights_single_CZ_htgsys.csv, revised_weights_CZ.csv and one excel file probably revised on the excel file received on May 17, 2014 from Vrushali: revised_weights_2014_june16_YLX.xlsx)
#   The following weights are taken from \\poncho\resstd\determination\IECC2015\quantitative\analysis\documentation\weights\weights.xlsx
#   

# eliminate all stuff
rm(list = ls(all = TRUE))

# close all devices which are currently open
device.list <- dev.list()
if (length(device.list) != 0){for (device.this in device.list){dev.off(device.this)}}

# setup start date and time, random seed
start_time <- date();
Start.time <- Sys.time()
set.seed(12345, kind = NULL)	# set seed of random number

# load libraries needed
library("reshape2")
library("lattice")
library("ggplot2")

source("multipleplot.R")

#
col.array <- rep(c("cyan","light green","green","grey","magenta","cyan","pink","black"),5)
col.array <- rep(c(rgb(68,118,178,maxColorValue=255),rgb(215,228,189,maxColorValue=255),"green","grey","magenta","cyan","pink","black"),5)

# today's month, day and year in the format of "Thu Jun 16 08:48:36 2011", 5 fields separated by space
today.month  <- strsplit(date(),"\\s+",perl=TRUE)[[1]][2]
today.day    <- strsplit(date(),"\\s+",perl=TRUE)[[1]][3]
today.year   <- strsplit(date(),"\\s+",perl=TRUE)[[1]][5]
today.hour   <- strsplit(strsplit(date(),"\\s+",perl=TRUE)[[1]][4],":",perl=TRUE)[[1]][1]
today.minute <- strsplit(strsplit(date(),"\\s+",perl=TRUE)[[1]][4],":",perl=TRUE)[[1]][2]
today.second <- strsplit(strsplit(date(),"\\s+",perl=TRUE)[[1]][4],":",perl=TRUE)[[1]][3]

# building areas
      footages  <- c(2400,1200)				# place holder: need to verify
names(footages) <- c("singlefamily","multifamily")


# -------------------------------------------------------------------------------------------------------------------------------------
# 	change to the script directory
# ------------------------------------------------------------------------------------------------------------------------------------- 
if(.Platform$OS.type == "unix") 
{
	Path.Current  <- "/phome/resstd/FOA/Phase1_Analysis/RCD_Analysis/0_scripts"
	Path.Project  <- "/phome/resstd/FOA/Phase1_Analysis/RCD_Analysis"
	Path.Simu     <- "/phome/resstd/FOA/Phase1_Analysis/RCD_Analysis/simulation/singlefamily"
}else{
	Path.Current  <- "Y:/FOA/Phase1_Analysis/RCD_Analysis/0_scripts"
	Path.Project  <- "Y:/FOA/Phase1_Analysis/RCD_Analysis"
	Path.Simu     <- "Y:/FOA/Phase1_Analysis/RCD_Analysis/simulation/singlefamily"
}

setwd(Path.Current)

# time stamp of data retrieved for the simulation and analysis
timeStamp.string <- "2015Jun15"

# arrays of states
array.states <- c("AL","AR","GA","KY","MD","NC","PA","TX")
	
# define log path and output weight folder
Path.LOG      <- paste(Path.Project,"0_log",sep="/")
Path.duct.TMP <- paste(Path.Project,"06A_duct_data_analysis",sep="/")
Path.duct.OUT <- paste(Path.duct.TMP,timeStamp.string,sep="/")
Path.Auxi     <- paste(Path.Project,"00_auxi_files",sep="/")
Path.share    <- paste(Path.Project,"04A_prepare_keyItem_for_simulation",timeStamp.string,sep="/")
if (!file.exists(Path.duct.TMP))  {print(paste("NOT existing:",Path.duct.TMP));dir.create(Path.duct.OUT,showWarnings=TRUE,recursive=TRUE)}
if (!file.exists(Path.duct.OUT))  {print(paste("NOT existing:",Path.duct.OUT));dir.create(Path.duct.OUT,showWarnings=TRUE,recursive=TRUE)}
if (!file.exists(Path.LOG))  {print(paste("NOT existing:",Path.LOG));dir.create(Path.LOG,showWarnings=TRUE,recursive=TRUE)}
if (!file.exists(Path.Auxi)) {print(paste("NOT existing:",Path.Auxi, " Check Why!",sep=""));die}
if (!file.exists(Path.share)){print(paste("NOT existing:",Path.share," Check Why!",sep=""));die}

FL.LOG <- paste(Path.LOG, "06A_duct_data_analysis.log",sep="/")
FL.EPW <- paste(Path.Auxi,"119_location_mapping_YLX.csv",sep="/")
if   (file.exists(FL.LOG)) {print(paste(FL.LOG,"exist.Delete it!"));file.remove(FL.LOG)}	
if (!(file.exists(FL.EPW))){print(paste("NOT existing:",FL.EPW," Check Why!",sep=""));die}	
cat(paste("1: define a log/output folder.\n",sep=""))
cat(paste("1: define a log/output folder.\n",sep=""),file=FL.LOG,append=TRUE)



# ---------------------------------------------------------------------------------------------------------
# mapping the city and climate zone: oad the weather file for the state: this is needed to map the city name and climate zone label
# ---------------------------------------------------------------------------------------------------------
myEPW.state <- read.table(FL.EPW,header=TRUE,stringsAsFactors=FALSE,sep=",")
cat(paste("2: weather file for the state are inputted.\n",sep=""))
cat(paste("2: weather file for the state are inputted.\n",sep=""),file=FL.LOG,append=TRUE)



# unit conversion
kWh_to_kbtu   <- 3.41214163 				# https://www.google.com/#q=kWh+to+btu&safe=off
therm_to_kbtu <- 100					# http://www.onlineconversion.com/energy.htm 	# therm_to_kbtu <- 99.976129 				# https://www.google.com/#q=therm+to+btu&safe=off
therm_to_Mbtu <- 0.1					# http://www.onlineconversion.com/energy.htm
Mbtu_to_therm <- 10	
cat(paste("3: unit conversion factors are specified.\n",sep=""))
cat(paste("3: unit conversion factors are specified.\n",sep=""),file=FL.LOG,append=TRUE)



# ---------------------------------------------------------------------------------------------------------
# shares
# ---------------------------------------------------------------------------------------------------------
# share of CZ in states: Sep 8, 2014 (Monday): Conference call changed the methodogy.  To meet deadline, 4 instead 6 CZ, 3 instead of 4 htgsys, 3 instead of foundation are included int he simulations
array.share.CZ.state <- data.frame(AL = c(0.181,0.819,0.000,0.000),
                                   AR = c(0.000,0.754,0.246,0.000),
                                   GA = c(0.197,0.726,0.077,0.000),
                                   KY = c(0.000,0.000,1.000,0.000),
                                   MD = c(0.000,0.000,1.000,0.000),
                                   NC = c(0.000,0.515,0.485,0.000),
                                   PA = c(0.000,0.000,0.234,0.766),
                                   TX = c(1.000,0.000,0.000,0.000))                      
row.names(array.share.CZ.state) <- c("Houston","Memphis","Baltimore","Chicago")
cat(paste("4: share of CZ in the state.  This will be replaced by share derived from field observations.\n",sep=""))
cat(paste("4: share of CZ in the state.  This will be replaced by share derived from field observations\n",sep=""),file=FL.LOG,append=TRUE)



# share of heating systems in states
array.share.htgSys.state <- data.frame(AL = c(0.689,0.289,0.000,0.021),
                                       AR = c(0.375,0.481,0.000,0.145),
                                       GA = c(0.789,0.190,0.001,0.020),
                                       KY = c(0.689,0.289,0.000,0.021),
                                       MD = c(0.789,0.190,0.001,0.020),
                                       NC = c(0.789,0.190,0.001,0.020),
                                       PA = c(0.245,0.692,0.046,0.017),
                                       TX = c(0.375,0.481,0.000,0.145))
row.names(array.share.htgSys.state) <- c("heatpump","gasfurnace","oilfurnace","electricfurnace")
cat(paste("5: share of htg system in the state.  This will be replaced by share derived from field observations\n",sep=""))
cat(paste("5: share of htg system in the state.  This will be replaced by share derived from field observations\n",sep=""),file=FL.LOG,append=TRUE)

   
# share of foundation in states
array.share.found.state <- data.frame(AL = c(0.441,0.086,0.106,0.367),
                                      AR = c(0.669,0.006,0.029,0.297),
                                      GA = c(0.571,0.066,0.097,0.267),
                                      KY = c(0.441,0.086,0.106,0.367),
                                      MD = c(0.280,0.307,0.183,0.230),
                                      NC = c(0.387,0.023,0.041,0.549),
                                      PA = c(0.289,0.246,0.328,0.137),
                                      TX = c(0.796,0.003,0.004,0.198))
row.names(array.share.found.state) <- c( "slab","heatedbsmt", "unheatedbsmt","crawlspace")
cat(paste("6: share of foundations in the state.  This will be replaced by share derived from field observations\n",sep=""))
cat(paste("6: share of foundations in the state.  This will be replaced by share derived from field observations\n",sep=""),file=FL.LOG,append=TRUE)


# normalize the share because they do not add up to 1
array.share.CZ.state     <- sweep(array.share.CZ.state,2,    colSums(array.share.CZ.state),    FUN="/")
array.share.htgSys.state <- sweep(array.share.htgSys.state,2,colSums(array.share.htgSys.state),FUN="/")
array.share.found.state  <- sweep(array.share.found.state,2, colSums(array.share.found.state), FUN="/")
cat(paste("7: normalize to 1 to the shares of CZ, htgSys and foundations in the state.\n",sep=""))
cat(paste("7: normalize to 1 to the shares of CZ, htgSys and foundations in the state.\n",sep=""),file=FL.LOG,append=TRUE)


cat(paste("8: start loopping through states.....................\n",sep=""))
cat(paste("8: start loopping through states.....................\n",sep=""),file=FL.LOG,append=TRUE)
for (this.state in array.states[1])
{
	# use the observation to update the CZ, htgSys and found shares in the states
	FL.found.htgSys.share.OBJ <- paste(Path.share,paste(paste("State",this.state,timeStamp.string,sep="_"),"_found.htgSys.share.Rdata",sep=""),sep="/")
	cat(paste("9A: file contains the shares derived from the field observations to be used to replace the shares used in the code development.\n",sep=""))
	cat(paste("9A: file contains the shares derived from the field observations to be used to replace the shares used in the code development.\n",sep=""),file=FL.LOG,append=TRUE)
	
	# delete existing data frames which will be loaded from [FL.found.htgSys.share.OBJ]
	if(exists("myTable.found.share")  && is.data.frame(get("myTable.found.share"))) {rm(myTable.found.share)}
	if(exists("myTable.fuel.share")   && is.data.frame(get("myTable.fuel.share")))  {rm(myTable.fuel.share)}
	if(exists("myTable.htgsys.share") && is.data.frame(get("myTable.htgsys.share"))){rm(myTable.htgsys.share)}
	if(exists("myTable.CZ.share")     && is.data.frame(get("myTable.CZ.share")))    {rm(myTable.CZ.share)}
	cat(paste("9B: deleted the possibly existsing shares tables before loading the field observation derived shares.\n",sep=""))
	cat(paste("9B: deleted the possibly existsing shares tables before loading the field observation derived shares.\n",sep=""),file=FL.LOG,append=TRUE)
	
	# load the observation distribution of CZ, htgsys and foundations, i.e., read in [myTable.CZ.share], [myTable.found.share] and [myTable.htgsys.share]
	load(file=FL.found.htgSys.share.OBJ)
	names(myTable.CZ.share)      <- sub("Var1","CZ",names(myTable.CZ.share))
	
	# manipulate the foundation shares
	names(myTable.CZ.share)         <- sub("Var1","CZ",sub("Freq",this.state,names(myTable.CZ.share)))
	myTable.CZ.share[,"CZ"]         <- gsub("^2A$","Houston",gsub("^3A$","Memphis",gsub("^4A$","Baltimore",gsub("^5A$","Chicago",myTable.CZ.share[,"CZ"]))))
	


	names(myTable.found.share)      <- sub("Var1","found",sub("Freq",this.state,names(myTable.found.share)))
	myTable.found.share[,"found"]   <- gsub("^UnheatedBasement$","unheatedbsmt",gsub("^Basement$","heatedbsmt",gsub("^Crawlspace$","crawlspace",gsub("^Slab$","slab",myTable.found.share[,"found"]))))




	names(myTable.htgsys.share)     <- sub("Var1","htgsys",sub("Freq",this.state,names(myTable.htgsys.share)))
	myTable.htgsys.share[,"htgsys"] <- gsub("^ElectricResistance$","electricfurnace",gsub("^HeatPump$","heatpump",gsub("^Furnace$","gasfurnace",myTable.htgsys.share[,"htgsys"])))
	cat(paste("9C: shares derived from field observations are uploaded and term are re-named to be consistent with the terms used in the previous shares tables.\n",sep=""))
	cat(paste("9C: shares derived from field observations are uploaded and term are re-named to be consistent with the terms used in the previous shares tables.\n",sep=""),file=FL.LOG,append=TRUE)

	# replace the foundation shares based on observations
	array.share.CZ.state.revised <- array.share.CZ.state
	array.share.CZ.state.revised["Houston",this.state]   <- ifelse(length(myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Houston",this.state]),  myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Houston",  this.state],0)
	array.share.CZ.state.revised["Memphis",this.state]   <- ifelse(length(myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Memphis",this.state]),  myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Memphis",  this.state],0)
	array.share.CZ.state.revised["Baltimore",this.state] <- ifelse(length(myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Baltimore",this.state]),myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Baltimore",this.state],0)
	array.share.CZ.state.revised["Chicago",this.state]   <- ifelse(length(myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Chicago",this.state]),  myTable.CZ.share[myTable.CZ.share[,"CZ"] == "Chicago",  this.state],0)
	array.share.CZ.state.revised  <- sweep(array.share.CZ.state.revised,2,colSums(array.share.CZ.state.revised),FUN="/")

	# replace The national CZ city name with the CZ label in [array.share.CZ.state.revised]
	row.names(array.share.CZ.state.revised) <- sub("Houston","2A",sub("Memphis","3A",sub("Baltimore","4A",sub("Chicago","5A",row.names(array.share.CZ.state.revised)))))
	
	# June 9, 2015: Change the CZ fraction based on Mark's number: 77.65% FOR 3A AND 22.35% FOR CZ 2A
	array.share.CZ.state.revised["3A","AL"] <- 0.7765
	array.share.CZ.state.revised["2A","AL"] <- 0.2235
	cat(paste("9D: Use CZ shares derived by Mark Halverson.\n",sep=""))
	cat(paste("9D: Use CZ shares derived by Mark Halverson.\n",sep=""),file=FL.LOG,append=TRUE)

	# replace the foundation shares based on observations
	array.share.found.state.revised <- array.share.found.state
	array.share.found.state.revised["slab",this.state]         <- ifelse(length(myTable.found.share[myTable.found.share[,"found"] == "slab",this.state]),        myTable.found.share[myTable.found.share[,"found"] == "slab",        this.state],0)
	array.share.found.state.revised["crawlspace",this.state]   <- ifelse(length(myTable.found.share[myTable.found.share[,"found"] == "crawlspace",this.state]),  myTable.found.share[myTable.found.share[,"found"] == "crawlspace",  this.state],0)
	array.share.found.state.revised["heatedbsmt",this.state]   <- ifelse(length(myTable.found.share[myTable.found.share[,"found"] == "heatedbsmt",this.state]),  myTable.found.share[myTable.found.share[,"found"] == "heatedbsmt",  this.state],0)
	array.share.found.state.revised["unheatedbsmt",this.state] <- ifelse(length(myTable.found.share[myTable.found.share[,"found"] == "unheatedbsmt",this.state]),myTable.found.share[myTable.found.share[,"found"] == "unheatedbsmt",this.state],0)
	array.share.found.state.revised  <- sweep(array.share.found.state.revised,2,colSums(array.share.found.state.revised),FUN="/")

	# replace the foundation shares based on observations
	array.share.htgSys.state.revised <- array.share.htgSys.state
	array.share.htgSys.state.revised["electricfurnace",this.state] <- ifelse(length(myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "electricfurnace",this.state]),myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "electricfurnace",this.state],0)
	array.share.htgSys.state.revised["gasfurnace",this.state]      <- ifelse(length(myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "gasfurnace",this.state]),     myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "gasfurnace",     this.state],0)
	array.share.htgSys.state.revised["heatpump",this.state]        <- ifelse(length(myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "heatpump",this.state]),       myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "heatpump",       this.state],0)
	array.share.htgSys.state.revised["oilfurnace",this.state]      <- ifelse(length(myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "oilfurnace",this.state]),     myTable.htgsys.share[myTable.htgsys.share[,"htgsys"] == "oilfurnace",     this.state],0)
	array.share.htgSys.state.revised  <- sweep(array.share.htgSys.state.revised,2,colSums(array.share.htgSys.state.revised),FUN="/")
	cat(paste("10: [",this.state,"]: the arrays for CZ, found and htgsys shares are updated based on field observation.\n",sep=""))
	cat(paste("10: [",this.state,"]: the arrays for CZ, found and htgsys shares are updated based on field observation.\n",sep=""),file=FL.LOG,append=TRUE)
	
	
	# ---------------------------------------------------------------------------------------------------------
	# from now on:  [array.share.CZ.state]     is replaced with [array.share.CZ.state.revised]
	#      		[array.share.found.state]  is replaced with [array.share.found.state.revised]
	#		[array.share.htgSys.state] is replaced with [array.share.htgSys.state.revised]
	# ---------------------------------------------------------------------------------------------------------
	
	
	# define sub-folder for each state
	Path.Out.State  <- paste(Path.duct.OUT, this.state,sep="/")
	# Path.Simu.State <- paste(Path.Simu,this.state,"duct.leakage",sep="/")
	Path.Simu.State <- paste(Path.Simu,"AL_June29_2015_Final","duct.leakage",sep="/")
	
	
	if (!file.exists(Path.Out.State)) {print(paste("NOT existing:",Path.Out.State)); dir.create(Path.Out.State, showWarnings=TRUE,recursive=TRUE)}
	if (!file.exists(Path.Simu.State)){print(paste("NOT existing:",Path.Simu.State,". Check Why!\n"));die}
	cat(paste("11: [",this.state,"]: specified an output and a simulation folder for the state.\n",sep=""))
	cat(paste("11: [",this.state,"]: specified an output and a simulation folder for the state.\n",sep=""),file=FL.LOG,append=TRUE)
	
	# define simulation subfolders
	Path.summary.nobackup <- paste(Path.Simu.State,"summary.nobackup",sep="/")
	if (!file.exists(Path.summary.nobackup)){print(paste("NOT existing:",Path.summary.nobackup),". Check Why!\n",sep="");die}
	
	# define E+ summary file for input
	FL.summary <- paste(Path.summary.nobackup,"summary.csv",sep="/")
	if (!file.exists(FL.summary)){print(paste("NOT existing:",FL.summary),". Check Why!\n",sep="");die}
	cat(paste("12: [",this.state,"]: define the E+ summary file name and path.\n",sep=""))
	cat(paste("12: [",this.state,"]: define the E+ summary file name and path.\n",sep=""),file=FL.LOG,append=TRUE)
	


	# define output file names
	FL.Rdata               <- paste(Path.Out.State,paste("FOA_data.Rdata",               sep=""),sep="/")
	FL.share.csv           <- paste(Path.Out.State,paste("FOA_Shares.csv",               sep=""),sep="/")	
	FL.RawData.csv         <- paste(Path.Out.State,paste("FOA_RawData.csv",              sep=""),sep="/")	
	FL.RawData.rds         <- paste(Path.Out.State,paste("FOA_RawData.rds",              sep=""),sep="/")	
	FL.Proc.Comp.csv       <- paste(Path.Out.State,paste("FOA_ProcData_Complete.csv",    sep=""),sep="/")	
	FL.Proc.Comp.rds       <- paste(Path.Out.State,paste("FOA_ProcData_Complete.rds",    sep=""),sep="/")	
	FL.Proc.Conc.csv       <- paste(Path.Out.State,paste("FOA_ProcData_Concise.csv",     sep=""),sep="/")	
	FL.Proc.Conc.rds       <- paste(Path.Out.State,paste("FOA_ProcData_Concise.rds",     sep=""),sep="/")	
	FL.Base.Advn.csv       <- paste(Path.Out.State,paste("FOA_Base_Advn.csv",            sep=""),sep="/")	
	FL.Base.Advn.obj       <- paste(Path.Out.State,paste("FOA_Base_Advn.Rdata",          sep=""),sep="/")	
	
	FL.EUI_dEUI.Raw.csv    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_RawModel.csv",    sep=""),sep="/")	
	FL.EUI_dEUI.Raw.rds    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_RawModel.rds",    sep=""),sep="/")	
	FL.EUI_dEUI.Raw.pdf    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_RawModel.pdf",    sep=""),sep="/")	
	
	FL.EUI_dEUI.Agg.csv    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_aggModel.csv",    sep=""),sep="/")	
	FL.EUI_dEUI.Agg.rds    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_aggModel.rds",    sep=""),sep="/")	
	FL.EUI_dEUI.Agg.pdf    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_aggModel.pdf",    sep=""),sep="/")	
	FL.EUI_dEUI.Agg.sum    <- paste(Path.Out.State,paste("FOA_EUI_dEUI_aggModel_sum.csv",sep=""),sep="/")
	
	FL.EUI_dEUI.rank.csv   <- paste(Path.Out.State,paste("FOA_EUI_dEUI_rankedModel.csv", sep=""),sep="/")
	
	FL.duct_EUI_change.csv <- paste(Path.Out.State,paste("FOA_duct_EUI_Change.csv",      sep=""),sep="/")
	FL.duct_EUI_change.rds <- paste(Path.Out.State,paste("FOA_duct_EUI_Change.rds",      sep=""),sep="/")
	FL.duct_EUI_Change.pdf <- paste(Path.Out.State,paste("FOA_duct_EUI_Change.pdf",      sep=""),sep="/")	
	
	if (file.exists(FL.Rdata           )){print(paste(FL.Rdata           ,    "exist.Delete it!"));file.remove(FL.Rdata           )}
	if (file.exists(FL.share.csv       )){print(paste(FL.share.csv       ,    "exist.Delete it!"));file.remove(FL.share.csv       )}
	if (file.exists(FL.RawData.csv     )){print(paste(FL.RawData.csv     ,    "exist.Delete it!"));file.remove(FL.RawData.csv     )}
	if (file.exists(FL.RawData.rds     )){print(paste(FL.RawData.rds     ,    "exist.Delete it!"));file.remove(FL.RawData.rds     )}
	if (file.exists(FL.Proc.Comp.csv   )){print(paste(FL.Proc.Comp.csv   ,    "exist.Delete it!"));file.remove(FL.Proc.Comp.csv   )}
	if (file.exists(FL.Proc.Comp.rds   )){print(paste(FL.Proc.Comp.rds   ,    "exist.Delete it!"));file.remove(FL.Proc.Comp.rds   )}
	if (file.exists(FL.Proc.Conc.csv   )){print(paste(FL.Proc.Conc.csv   ,    "exist.Delete it!"));file.remove(FL.Proc.Conc.csv   )}
	if (file.exists(FL.Proc.Conc.rds   )){print(paste(FL.Proc.Conc.rds   ,    "exist.Delete it!"));file.remove(FL.Proc.Conc.rds   )}
	if (file.exists(FL.Base.Advn.csv   )){print(paste(FL.Base.Advn.csv   ,    "exist.Delete it!"));file.remove(FL.Base.Advn.csv   )}
	if (file.exists(FL.Base.Advn.obj   )){print(paste(FL.Base.Advn.obj   ,    "exist.Delete it!"));file.remove(FL.Base.Advn.obj   )}
	if (file.exists(FL.EUI_dEUI.Raw.csv)){print(paste(FL.EUI_dEUI.Raw.csv,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Raw.csv)}
	if (file.exists(FL.EUI_dEUI.Raw.rds)){print(paste(FL.EUI_dEUI.Raw.rds,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Raw.rds)}
	if (file.exists(FL.EUI_dEUI.Raw.pdf)){print(paste(FL.EUI_dEUI.Raw.pdf,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Raw.pdf)}
	if (file.exists(FL.EUI_dEUI.Agg.csv)){print(paste(FL.EUI_dEUI.Agg.csv,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Agg.csv)}
	if (file.exists(FL.EUI_dEUI.Agg.rds)){print(paste(FL.EUI_dEUI.Agg.rds,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Agg.rds)}
	if (file.exists(FL.EUI_dEUI.Agg.pdf)){print(paste(FL.EUI_dEUI.Agg.pdf,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Agg.pdf)}
	if (file.exists(FL.EUI_dEUI.Agg.sum)){print(paste(FL.EUI_dEUI.Agg.sum,    "exist.Delete it!"));file.remove(FL.EUI_dEUI.Agg.sum)}
	
	if (file.exists(FL.EUI_dEUI.rank.csv)){print(paste(FL.EUI_dEUI.rank.csv,  "exist.Delete it!"));file.remove(FL.EUI_dEUI.rank.csv)}

	if (file.exists(FL.duct_EUI_change.csv)){print(paste(FL.duct_EUI_change.csv,  "exist.Delete it!"));file.remove(FL.duct_EUI_change.csv)}
	if (file.exists(FL.duct_EUI_change.rds)){print(paste(FL.duct_EUI_change.rds,  "exist.Delete it!"));file.remove(FL.duct_EUI_change.rds)}
	if (file.exists(FL.duct_EUI_Change.pdf)){print(paste(FL.duct_EUI_Change.pdf,  "exist.Delete it!"));file.remove(FL.duct_EUI_Change.pdf)}

	
	cat(paste("13: [",this.state,"]: define output files.\n",sep=""))
	cat(paste("13: [",this.state,"]: define output files.\n",sep=""),file=FL.LOG,append=TRUE)

	# open pdf file fr plotting
	pdf(file = FL.EUI_dEUI.Raw.pdf,   paper="special", width=17, height=11,bg = "transparent")	
	pdf(file = FL.EUI_dEUI.Agg.pdf,   paper="special", width=17, height=11,bg = "transparent")	
	pdf(file = FL.duct_EUI_Change.pdf,paper="special", width=17, height=11,bg = "transparent")	
	
	# ---------------------------------------------------------------------------
	# output both the general shares used in the code development and the fiel observation derived share from this FOA analysis
	# ---------------------------------------------------------------------------
	cat("\narray.share.htgSys.state (general),",file=FL.share.csv,append=TRUE)
	write.table(array.share.htgSys.state,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	cat("\narray.share.htgSys.state.revised (from the FOA survey),",file=FL.share.csv,append=TRUE)
	write.table(array.share.htgSys.state.revised,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)

	cat("\narray.share.found.state (general),",file=FL.share.csv,append=TRUE)
	write.table(array.share.found.state,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	cat("\narray.share.found.state.revised (from the FOA survey),",file=FL.share.csv,append=TRUE)
	write.table(array.share.found.state.revised,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)

	cat("\narray.share.CZ.state (general),",file=FL.share.csv,append=TRUE)
	write.table(array.share.CZ.state,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	cat("\narray.share.CZ.state.revised (from the FOA survey),",file=FL.share.csv,append=TRUE)
	write.table(array.share.CZ.state.revised,file=FL.share.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)	
	cat(paste("14: [",this.state,"]: output the shares used in the analysis for verification purpose: [array.share.htgSys.state.revised], [array.share.found.state.revised] and [array.share.CZ.state.revised].\n",sep=""))
	cat(paste("14: [",this.state,"]: output the shares used in the analysis for verification purpose: [array.share.htgSys.state.revised], [array.share.found.state.revised] and [array.share.CZ.state.revised].\n",sep=""),file=FL.LOG,append=TRUE)




	# **********************************************************************************
	# field name: NEED VERIFICATION!!!!!: this needs changes when the new 15 IECC Climate Cities are ready!!!!!!!!!!!!!!
	# [fields.raw.common]: the fields in the summary file
	# [fields.common]:     the fields renamed (some "^S\.":	field needed to be keep
	#                                      and some "SS.":  fields associated with the original 9 key code items such as ach50 etc
	# ********************************************************************************** 
 	fields.raw.common <- c("idx",  "ID","Total.Site.Energy.kBtu","Site.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm",  "Heating.kWh",  "Heating.therm",  "Cooling.kWh",  "Cooling.therm",  "Interior.Lighting.kWh",  "Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm",  "Water.Systems.kWh",  "Water.Systems.therm",  "Fan_CoolingOp",      "Fan_HeatingOp",    "Location",   "ach50","afn_control","aspect_ratio","cfa_total",  "code",  "dhw_type",  "doe_climate_zone",  "doe_moisture_regime",   "f_inc_hw",  "fndn_type",  "heating_fuel",                "leakage_ratio",  "location","n_stories_per_unit","n_units",  "prototype",   "r_bsmtwall",   "r_bsmtwall_raw",   "r_ceiling",   "r_ceiling_raw",   "r_floor",   "r_floor_raw",   "r_sheathing",   "r_slab",   "r_wall",   "r_wall_raw",                   "runbsmt",  "runslab",   "shgc_window",  "system",  "tag",   "u_window",  "vent_sch",  "weatherfile",  "wfr","  cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw")
	fields.common     <- c("idx","S.ID","Total.Site.Energy.kBtu","Site.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm","S.Heating.kWh","S.Heating.therm","S.Cooling.kWh","S.Cooling.therm","S.Interior.Lighting.kWh","S.Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm","S.Water.Systems.kWh","S.Water.Systems.therm","S.Fan_CoolingOp.kWh","S.Fan_HeatingOp.kWh","Location","SS.ach50","afn_control","aspect_ratio","cfa_total","S.code","S.dhw_type","S.doe_climate_zone","S.doe_moisture_regime","SS.f_inc_hw","S.fndn_type","S.heating_fuel",             "SS.leakage_ratio","S.Location","n_stories_per_unit","n_units","S.prototype","SS.r_bsmtwall","SS.r_bsmtwall_raw","SS.r_ceiling","SS.r_ceiling_raw","SS.r_floor","SS.r_floor_raw","SS.r_sheathing","SS.r_slab","SS.r_wall","SS.r_wall_raw",                 "S.runbsmt","S.runslab","SS.shgc_window","S.system","S.tag","SS.u_window","S.vent_sch","S.weatherfile","S.wfr","S.cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw") 	
#	                        ID	     Total.Site.Energy.kBtu	 Site.EUI.kBtu.sqft	 Total.Electricity.kWh	 Total.Natural.Gas.therm     Heating.kWh     Heating.therm     Cooling.kWh     Cooling.therm	 Interior.Lighting.kWh	   Interior.Lighting.therm   Interior.Equipment.kWh   Interior.Equipment.therm	 Fans.kWh   Fans.therm	   Water.Systems.kWh	 Water.Systems.therm	 Fan_CoolingOp	       Fan_HeatingOp	   Location	         afn_control   aspect_ratio   cfa_total	    code     dhw_type	  doe_climate_zone     doe_moisture_regime                   fndn_type	   heating_fuel	identifier	  leakage_ratio	    location   n_stories_per_unit   n_units     prototype	                                                                                                                                                           r_returnduct	    runbsmt     runslab	                     system     tag	              vent_sch	   weatherfile	   wfr	   cfa_per_unit	  ela   f_cfl_hw   f_inc_hw     f_led_hw   f_lf_hw
 	fields.raw.common <- c("idx",  "ID","Total.Site.Energy.kBtu","Site.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm",  "Heating.kWh",  "Heating.therm",  "Cooling.kWh",  "Cooling.therm",  "Interior.Lighting.kWh",  "Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm",  "Water.Systems.kWh",  "Water.Systems.therm",  "Fan_CoolingOp",      "Fan_HeatingOp",    "Location",           "afn_control","aspect_ratio","cfa_total",  "code",  "dhw_type",  "doe_climate_zone",  "doe_moisture_regime",                "fndn_type",  "heating_fuel","identifier",   "leakage_ratio",  "location","n_stories_per_unit","n_units",  "prototype",                                                                                                                                                               "r_returnduct",  "runbsmt",  "runslab",                   "system",  "tag",                "vent_sch",  "weatherfile",  "wfr","  cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw")
 	fields.common     <- c("idx","S.ID","Total.Site.Energy.kBtu","Site.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm","S.Heating.kWh","S.Heating.therm","S.Cooling.kWh","S.Cooling.therm","S.Interior.Lighting.kWh","S.Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm","S.Water.Systems.kWh","S.Water.Systems.therm","S.Fan_CoolingOp.kWh","S.Fan_HeatingOp.kWh","Location",           "afn_control","aspect_ratio","cfa_total","S.code","S.dhw_type","S.doe_climate_zone","S.doe_moisture_regime",              "S.fndn_type","S.heating_fuel","identifier","SS.leakage_ratio","S.Location","n_stories_per_unit","n_units","S.prototype",                                                                                                                                                               "r_returnduct","S.runbsmt","S.runslab",                 "S.system","S.tag",              "S.vent_sch","S.weatherfile","S.wfr","S.cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw") 	

	# June 29, 2015: the output fileds changes due to recent template modification.
	fields.raw.common <- c("idx",  "ID","Total.Site.Energy.kBtu","Total.Source.Energy.kBtu",         "Site.EUI.kBtu.sqft",       "Source.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm",  "Heating.kWh",  "Heating.therm",  "Cooling.kWh",  "Cooling.therm",  "Interior.Lighting.kWh",  "Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm",  "Water.Systems.kWh",  "Water.Systems.therm",  "Fan_CoolingOp",      "Fan_HeatingOp",    "Location",           "afn_control","aspect_ratio","cfa_total",  "code",  "dhw_type",  "doe_climate_zone",  "doe_moisture_regime",                "fndn_type",  "heating_fuel","identifier",   "leakage_ratio",  "location","n_stories_per_unit","n_units",  "prototype",                                                                                                                                                               "r_returnduct",  "runbsmt",  "runslab",                   "system",  "tag",                "vent_sch",  "weatherfile",  "wfr","  cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw")
	fields.common     <- c("idx","S.ID","Total.Site.Energy.kBtu","Total.Source.Energy.kBtu",         "Site.EUI.kBtu.sqft",       "Source.EUI.kBtu.sqft","Total.Electricity.kWh","Total.Natural.Gas.therm","S.Heating.kWh","S.Heating.therm","S.Cooling.kWh","S.Cooling.therm","S.Interior.Lighting.kWh","S.Interior.Lighting.therm","Interior.Equipment.kWh","Interior.Equipment.therm","Fans.kWh","Fans.therm","S.Water.Systems.kWh","S.Water.Systems.therm","S.Fan_CoolingOp.kWh","S.Fan_HeatingOp.kWh","Location",           "afn_control","aspect_ratio","cfa_total","S.code","S.dhw_type","S.doe_climate_zone","S.doe_moisture_regime",              "S.fndn_type","S.heating_fuel","identifier","SS.leakage_ratio","S.Location","n_stories_per_unit","n_units","S.prototype",                                                                                                                                                               "r_returnduct","S.runbsmt","S.runslab",                 "S.system","S.tag",              "S.vent_sch","S.weatherfile","S.wfr","S.cfa_per_unit","ela","f_cfl_hw","f_inc_hw.1","f_led_hw","f_lf_hw") 		
	cat(paste("15: [",this.state,"]: ensure the field names are correct!!!! Note: the fields are changed since Phase 0 sample design simulation.\n",sep=""))
	cat(paste("15: [",this.state,"]: ensure the field names are correct!!!! Note: the fields are changed since Phase 0 sample design simulation.\n",sep=""),file=FL.LOG,append=TRUE)

	# ---------------------------------------------------------------------------------
	# read simulation summary data into [myData.raw]
	# ---------------------------------------------------------------------------------
	myData.raw <- read.table(file=FL.summary,header=TRUE,sep=",",stringsAsFactors=FALSE)
	cat(paste("16: [",this.state,"]: read the summary file from simulation.\n",sep=""))
	cat(paste("16: [",this.state,"]: read the summary file from simulation.\n",sep=""),file=FL.LOG,append=TRUE)


	# change field names
	names(myData.raw) <- fields.common
	cat(paste("17: [",this.state,"]: change the field names.\n",sep=""))
	cat(paste("17: [",this.state,"]: change the field names.\n",sep=""),file=FL.LOG,append=TRUE)

	# ----------------------------------------------------------------------------------
	# output [myData.raw]:	immediate output the data read in with just the field name change
	# ----------------------------------------------------------------------------------
	cat(paste("raw data,",sep=""),file=FL.RawData.csv,append=TRUE)
	write.table(myData.raw,file=FL.RawData.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
		
	saveRDS(myData.raw,file=FL.RawData.rds)
	cat(paste("18: [",this.state,"]: output [myData.raw] to [",FL.RawData.rds,"] before adding any new field.\n",sep=""))
	cat(paste("18: [",this.state,"]: output [myData.raw] to [",FL.RawData.rds,"] before adding any new field.\n",sep=""),file=FL.LOG,append=TRUE)

















	# ---------------------------------------------------------------------------------
	# append some auxilary fields 
	# ---------------------------------------------------------------------------------

	# added a "S.fuel" field which expand the two fuels in "S.heating_fuel" field into three by assign "oil" to "oilfurnace" of the "S.heating_fuel" field.
	# "S.heating_fuel" does not distinguish "natural gas" and "oil", add a "fuel" field based on "S.tag" field"
	# change "naturalgas" to "oil" when AC Furnace use oil as fuel for heating
	myData.raw[,"S.fuel"] <- myData.raw[,"S.heating_fuel"]
	myData.raw[myData.raw[,"S.tag"]=="oilfurnace","S.fuel"] <- "oil"
	cat(paste("19A: [",this.state,"]: create a [S.fuel] field with [oil] assigned to oilfurnace.\n",sep=""))
	cat(paste("19A: [",this.state,"]: create a [S.fuel] field with [oil] assigned to oilfurnace.\n",sep=""),file=FL.LOG,append=TRUE)

	# added a "S.htgSys" for heating system based on "S.Tag", which has 4 otions i.e., "HeatPump", "ElecResis", "NaturalGas", "LPG"
	myData.raw[,"S.htgSys"]                                        <- "HeatPump"	# initialize the "S.htgSys" field with "HeatPump"
	myData.raw[myData.raw[,"S.tag"]=="gasfurnace",     "S.htgSys"] <- "NaturalGas"	# 
	myData.raw[myData.raw[,"S.tag"]=="oilfurnace",     "S.htgSys"] <- "LPG"
	myData.raw[myData.raw[,"S.tag"]=="electricfurnace","S.htgSys"] <- "ElecResis"
	cat(paste("19B: [",this.state,"]: added a \"S.htgSys\" for heating system based on \"S.Tag\", which has 4 options i.e., \"HeatPump\", \"ElecResis\", \"NaturalGas\", \"LPG\".\n",sep=""))
	cat(paste("19B: [",this.state,"]: added a \"S.htgSys\" for heating system based on \"S.Tag\", which has 4 options i.e., \"HeatPump\", \"ElecResis\", \"NaturalGas\", \"LPG\".\n",sep=""),file=FL.LOG,append=TRUE)

	# ---------------------------------------------------------------------------------
	# from [myData.raw] to [myData.Proce]: further extract only those fields started with "S." into [myData]
	# ---------------------------------------------------------------------------------
	myData.Proc <- myData.raw[,grep("S\\.",names(myData.raw),value=TRUE)]		# this includes both "S." and "SS." fields	
	cat(paste("20A: [",this.state,"]: [myData.raw] into [myData.Proc] but only keep fields started with S. or SS.\n",sep=""))
	cat(paste("20A: [",this.state,"]: [myData.raw] into [myData.Proc] but only keep fields started with S. or SS.\n",sep=""),file=FL.LOG,append=TRUE)

	# add an footage field to [myData.Proc]
	myData.Proc[,"S.area"] <- footages[myData.Proc[,"S.prototype"]]
	cat(paste("20B: [",this.state,"]: added an [S.area] field in [myData.Proc].\n",sep=""))
	cat(paste("20B: [",this.state,"]: added an [S.area] field in [myData.Proc].\n",sep=""),file=FL.LOG,append=TRUE)




	# ---------------------------------------------------------------------------------
	# add ["S.htg.kWh],   ["S.clg.kWh],   ["S.ltg.kWh]   and ["S.swh.kWh] AND     
	#     ["S.htg.therm], ["S.clg.therm], ["S.ltg.therm] and ["S.swh.therm] fields
	# ---------------------------------------------------------------------------------
	# Process Regulated Loads: Put the processed fields into "SS." fields
	#
	# Create "S.htg.kWh",  "S.clg.kWh",  "S.ltg.kWh",  "S.swh.kWh"   after including the fan portion for cooling and heating
	# Create "S.htg.therm","S.clg.therm","S.ltg.therm","S.swh.therm" 
	#
	# create "SS." series of variables and add the fan energy to the heating and cooling energy in [myData.Proc]
	#
	# Convert "S.Heating.kWh" + "S.Fan_HeatingOp.kWh" --> "S.htg.kWh"
	# Convert "S.Cooling.kWh" + "S.Fan_CoolingOp.kWh" --> "S.clg.kWh"
	# Convert "S.Interior.Lighting.kWh"               --> "S.ltg.kWh"
	# Convert "S.Water.Systems.kWh"                   --> "S.swh.kWh"
	#
	# Convert "S.Heating.therm"                       --> "S.htg.therm"
	# Convert "S.Cooling.therm"                       --> "S.clg.therm"
	# Convert "S.Interior.Lighting.therm"             --> "S.ltg.therm"
	# Convert "S.Water.Systems.therm"                 --> "S.swh.therm"
	# ---------------------------------------------------------------------------------
	myData.Proc[,"S.htg.kWh"]   <- myData.Proc[,"S.Heating.kWh"] + myData.Proc[,"S.Fan_HeatingOp.kWh"] 	
	myData.Proc[,"S.clg.kWh"]   <- myData.Proc[,"S.Cooling.kWh"] + myData.Proc[,"S.Fan_CoolingOp.kWh"] 	
	myData.Proc[,"S.ltg.kWh"]   <- myData.Proc[,"S.Interior.Lighting.kWh"] 					
	myData.Proc[,"S.swh.kWh"]   <- myData.Proc[,"S.Water.Systems.kWh"] 					

	myData.Proc[,"S.htg.therm"] <- myData.Proc[,"S.Heating.therm"] 						
	myData.Proc[,"S.clg.therm"] <- myData.Proc[,"S.Cooling.therm"] 						
	myData.Proc[,"S.ltg.therm"] <- myData.Proc[,"S.Interior.Lighting.therm"] 				
	myData.Proc[,"S.swh.therm"] <- myData.Proc[,"S.Water.Systems.therm"] 					
	cat(paste("21: [",this.state,"]: fan energy added to htg/clg energy and a new series of variable starting with [S.] is created.\n",sep=""))
	cat(paste("21: [",this.state,"]: fan energy added to htg/clg energy and a new series of variable starting with [S.] is created.\n",sep=""),file=FL.LOG,append=TRUE)




	# ---------------------------------------------------------------------------------
	#
	# enduse EUI: 
	#
	# add ["S.htg.EUI.elec],["S.clg.EUI.elec],["S.ltg.EUI.elec] and ["S.swh.EUI.elec] AND     
	#     ["S.htg.EUI.gas], ["S.clg.EUI.gas], ["S.ltg.EUI.gas]  and ["S.swh.EUI.gas]  AND
	#     ["S.htg.EUI.sum], ["S.clg.EUI.sum], ["S.ltg.EUI.sum]  and ["S.swh.EUI.sum] fields	
	# ---------------------------------------------------------------------------------	
	# create EUI fields: combine the energy usage of both "kWh" from electricity and "therm" from both "natural gas" and "LPG"
	myData.Proc[,"S.htg.EUI.elec"] <- (myData.Proc[,"S.htg.kWh"] * kWh_to_kbtu) / myData.Proc[,"S.area"]
	myData.Proc[,"S.clg.EUI.elec"] <- (myData.Proc[,"S.clg.kWh"] * kWh_to_kbtu) / myData.Proc[,"S.area"] 	
	myData.Proc[,"S.ltg.EUI.elec"] <- (myData.Proc[,"S.ltg.kWh"] * kWh_to_kbtu) / myData.Proc[,"S.area"]
	myData.Proc[,"S.swh.EUI.elec"] <- (myData.Proc[,"S.swh.kWh"] * kWh_to_kbtu) / myData.Proc[,"S.area"]

	myData.Proc[,"S.htg.EUI.gas"]  <- (myData.Proc[,"S.htg.therm"] * therm_to_kbtu) / myData.Proc[,"S.area"]
	myData.Proc[,"S.clg.EUI.gas"]  <- (myData.Proc[,"S.clg.therm"] * therm_to_kbtu) / myData.Proc[,"S.area"] 	
	myData.Proc[,"S.ltg.EUI.gas"]  <- (myData.Proc[,"S.ltg.therm"] * therm_to_kbtu) / myData.Proc[,"S.area"]
	myData.Proc[,"S.swh.EUI.gas"]  <- (myData.Proc[,"S.swh.therm"] * therm_to_kbtu) / myData.Proc[,"S.area"]

	myData.Proc[,"S.htg.EUI.sum"]  <- myData.Proc[,"S.htg.EUI.elec"] + myData.Proc[,"S.htg.EUI.gas"]
	myData.Proc[,"S.clg.EUI.sum"]  <- myData.Proc[,"S.clg.EUI.elec"] + myData.Proc[,"S.clg.EUI.gas"]	
	myData.Proc[,"S.ltg.EUI.sum"]  <- myData.Proc[,"S.ltg.EUI.elec"] + myData.Proc[,"S.ltg.EUI.gas"]
	myData.Proc[,"S.swh.EUI.sum"]  <- myData.Proc[,"S.swh.EUI.elec"] + myData.Proc[,"S.swh.EUI.gas"]
	cat(paste("22: [",this.state,"]: create EUI fields.\n",sep=""))
	cat(paste("22: [",this.state,"]: create EUI fields.\n",sep=""),file=FL.LOG,append=TRUE)



	# ---------------------------------------------------------------------------------
	#
	# IECC Regulated end-use EUI (sum of enduse EUIs):
	#
	# add ["SS.tot.EUI.elec"] <- ["S.tot.EUI.elec"] <- sum(["S.htg.EUI.elec],["S.clg.EUI.elec],["S.ltg.EUI.elec] and ["S.swh.EUI.elec]) AND     
	#     ["SS.tot.EUI.gas"]  <- ["S.tot.EUI.gas"]  <- sum(["S.htg.EUI.gas], ["S.clg.EUI.gas], ["S.ltg.EUI.gas]  and ["S.swh.EUI.gas])  AND
	#     ["SS.tot.EUI.sum"]  <- ["S.tot.EUI.sum"]  <- sum(["S.htg.EUI.sum], ["S.clg.EUI.sum], ["S.ltg.EUI.sum]  and ["S.swh.EUI.sum]) fields	
	# ---------------------------------------------------------------------------------
	myData.Proc[,"S.tot.EUI.elec"]  <- myData.Proc[,"S.htg.EUI.elec"]  + myData.Proc[,"S.clg.EUI.elec"]  + myData.Proc[,"S.ltg.EUI.elec"]  + myData.Proc[,"S.swh.EUI.elec"] 
	myData.Proc[,"S.tot.EUI.gas"]   <- myData.Proc[,"S.htg.EUI.gas"]   + myData.Proc[,"S.clg.EUI.gas"]   + myData.Proc[,"S.ltg.EUI.gas"]   + myData.Proc[,"S.swh.EUI.gas"] 
	myData.Proc[,"S.tot.EUI.sum"]   <- myData.Proc[,"S.htg.EUI.sum"]   + myData.Proc[,"S.clg.EUI.sum"]   + myData.Proc[,"S.ltg.EUI.sum"]   + myData.Proc[,"S.swh.EUI.sum"] 

	myData.Proc[,"SS.tot.EUI.elec"]  <- myData.Proc[,"S.tot.EUI.elec"] 
	myData.Proc[,"SS.tot.EUI.gas"]   <- myData.Proc[,"S.tot.EUI.gas"]  
	myData.Proc[,"SS.tot.EUI.sum"]   <- myData.Proc[,"S.tot.EUI.sum"]  
	cat(paste("23: [",this.state,"]: summing the individual enduse EUI/Cost to a total EUI/Cost.\n",sep=""))
	cat(paste("23: [",this.state,"]: summing the individual enduse EUI/Cost to a total EUI/Cost.\n",sep=""),file=FL.LOG,append=TRUE)


	# ---------------------------------------------------------------------------------
	# break up the "S.ID" to get auxilary fields we need: NOTE: this need to be checked with the actual case name convention used!!!!!!!!!!!!
	# ---------------------------------------------------------------------------------
	myData.Proc[,"SS.type"]       <- sub("(.*)_(.*)_(.*)_(.*)_(.*)","\\1",myData.Proc[,"S.ID"],perl=TRUE)			
	myData.Proc[,"SS.City.CZ"]    <- sub("(.*)_(.*)_(.*)_(.*)_(.*)","\\2",myData.Proc[,"S.ID"],perl=TRUE)			
	myData.Proc[,"SS.Standard"]   <- sub("(.*)_(.*)_(.*)_(.*)_(.*)","\\3",myData.Proc[,"S.ID"],perl=TRUE)			
	tmp1                          <- sub("(.*)_(.*)_(.*)_(.*)_(.*)","\\4",myData.Proc[,"S.ID"],perl=TRUE)
	myData.Proc[,"SS.PRO.IDX"]    <- sub("(.*)_(.*)_(.*)_(.*)_(.*)","\\5",myData.Proc[,"S.ID"],perl=TRUE)
	myData.Proc[,"SS.HeatSys"]    <- sub("(.*)\\.(.*)",             "\\1",tmp1,                perl=TRUE)
	myData.Proc[,"SS.Foundation"] <- sub("(.*)\\.(.*)",             "\\2",tmp1,                perl=TRUE)


	myData.Proc[,"SS.City"]       <- sub("(.*)\\.(.*)","\\1",myData.Proc[,"SS.City.CZ"],perl=TRUE)
	myData.Proc[,"SS.CZ"]         <- sub("(.*)\\.(.*)","\\2",myData.Proc[,"SS.City.CZ"],perl=TRUE)
	myData.Proc[,"SS.CZ.label"]   <- myEPW.state[match(myData.Proc[,"SS.City"],myEPW.state[,"CZ.city"]),"CZ.revised"]
	cat(paste("24: [",this.state,"]: split htgSys, foundation, code item evaluated from the case ID.\n",sep=""))
	cat(paste("24: [",this.state,"]: split htgSys, foundation, code item evaluated from the case ID.\n",sep=""),file=FL.LOG,append=TRUE)


	# --------------------------------------------------------------------------------- 
	# add a SS.ID based on the combination of "SS.City","SS.HeatSys","SS.Foundation"
	# ---------------------------------------------------------------------------------
	myData.Proc[,"SS.ID"] <- paste(paste(myData.Proc[,"SS.City"],myData.Proc[,"SS.CZ"],sep=""),myData.Proc[,"SS.HeatSys"],myData.Proc[,"SS.Foundation"],sep="_")
	cat(paste("25: [",this.state,"]: add a ID based on combination of city, htgsys and foundation.\n",sep=""))
	cat(paste("25: [",this.state,"]: add a ID based on combination of city, htgsys and foundation.\n",sep=""),file=FL.LOG,append=TRUE)


	# ----------------------------------------------------------------------------------
	# OUTPUTTING THE DATA BEFORE AGGREGATION
	# output [myData.Proc]:	output the processed data before aggregating across heating systems and foundations
	# wirte to [FOA_ProcData_Complete.csv] and [FOA_ProcData_Complete.rds]
	# ----------------------------------------------------------------------------------
	cat(paste("summary data,",sep=""),file=FL.Proc.Comp.csv,append=TRUE)
	write.table(myData.Proc,file=FL.Proc.Comp.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	cat(paste("\n\n",sep=""),file=FL.Proc.Comp.csv,append=TRUE)
	
	saveRDS(myData.Proc,file=FL.Proc.Comp.rds)
	cat(paste("26: [",this.state,"]: output [myData.Proc] to [",FL.Proc.Comp.rds,"] and [",FL.Proc.Comp.csv,"].\n",sep=""))
	cat(paste("26: [",this.state,"]: output [myData.Proc] to [",FL.Proc.Comp.rds,"] and [",FL.Proc.Comp.csv,"].\n",sep=""),file=FL.LOG,append=TRUE)



	# ---------------------------------------------------------------------------------
	# OUTPUTTING THE DATA BEFORE AGGREGATION
	# from [myData.Proc] to [myData.Concise] by only keeping the "SS." fields
	# write to [FOA_ProcData_Concise.csv] and [FOA_ProcData_Concise.rds] fields
	# ---------------------------------------------------------------------------------
	myData.Concise <- myData.Proc[,grep("^SS\\.",names(myData.Proc),value=TRUE)]
	cat(paste("27: [",this.state,"]: only keep essential fields of [myData.Proc] in [myData.Concise].\n",sep=""))
	cat(paste("27: [",this.state,"]: only keep essential fields of [myData.Proc] in [myData.Concise].\n",sep=""),file=FL.LOG,append=TRUE)

	# ----------------------------------------------------------------------------------
	# output [myData.Concise]:	
	# ----------------------------------------------------------------------------------
	cat(paste("Concise data,",sep=""),file=FL.Proc.Conc.csv,append=TRUE)
	write.table(myData.Concise,file=FL.Proc.Conc.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	saveRDS(myData.Concise,file=FL.Proc.Conc.rds)
	cat(paste("28: [",this.state,"]: output [myData.Concise].\n",sep=""))
	cat(paste("28: [",this.state,"]: output [myData.Concise].\n",sep=""),file=FL.LOG,append=TRUE)



	# ---------------------------------------------------------------------------------
	# from [myData.Concise] to [myData.DIF] ([myData.PRO] and [myData.TRB]
	# ---------------------------------------------------------------------------------
	# subset TRB and PRO data into [myData.PRO] and [myData.TRB]
	myData.PRO <- myData.Concise[grep("PRO",myData.Concise[,"SS.type"]),]
	myData.TRB <- myData.Concise[grep("TRB",myData.Concise[,"SS.type"]),]
	cat(paste("29: [",this.state,"]: Separate PRO and TRB data of [myData.Concise] into ",dim(myData.PRO)[1]," by ",dim(myData.PRO)[2]," [myData.PRO] and ",dim(myData.TRB)[1]," by ",dim(myData.TRB)[2]," [myData.TRB].\n",sep=""))
	cat(paste("29: [",this.state,"]: Separate PRO and TRB data of [myData.Concise] into ",dim(myData.PRO)[1]," by ",dim(myData.PRO)[2]," [myData.PRO] and ",dim(myData.TRB)[1]," by ",dim(myData.TRB)[2]," [myData.TRB].\n",sep=""),file=FL.LOG,append=TRUE)


	# ----------------------------------------------------------------------------------
	#                               [Part.Base] <- [myData.BASE] <- [myData.TRB] <- myData.Concise]
	#                               [Part.Advn] <- [myData.ADVN] <- [myData.PRO] <- myData.Concise]
	#                [Part.Diff] <- [Part.Advn] & [Part.Base]
	# [myEUI.raw] <- [Part.Base] & [Part.Advn] & [Part.Diff]
	#
	# NOTE: the fields to be retained (fields.chr) (subsequently (fields.num) are manually specified here!!!!!!!!!!!!
	# calculating [myEUI.raw] which contains the difference between the PRO and their corresponding TRB case
	# ----------------------------------------------------------------------------------
	myData.BASE    <- myData.TRB[match(myData.PRO[,"SS.ID"],myData.TRB[,"SS.ID"]),]	 								# make a array the same size as [myData.PRO] but filled with corresponding value in [myData.TRB]
	fields.chr     <- c("SS.type","SS.City.CZ","SS.Standard","SS.PRO.IDX","SS.HeatSys","SS.Foundation","SS.City","SS.CZ","SS.CZ.label","SS.ID")	# NOTE: this list may vary.  It needs to be checked.  Basically it consists of all fields other than the code items fields.
	fields.num     <- names(myData.PRO)[!(names(myData.PRO) %in% fields.chr)]
	myData.ADVN    <- myData.PRO[,fields.num] 
	       tmp     <- myData.ADVN[,fields.num] - myData.BASE[,fields.num]
	       tmp.pct <- 100 * (myData.ADVN[,fields.num] - myData.BASE[,fields.num]) / myData.BASE[,fields.num]
	       
	Part.Advn  <- cbind(myData.PRO[,fields.chr], myData.PRO[,fields.num])
  names(Part.Advn) <- paste("PRO",names(Part.Advn),sep="_")
  
	Part.Base <- cbind(myData.BASE[,fields.chr],myData.BASE[,fields.num])
  names(Part.Base) <- paste("TRB",names(Part.Base),sep="_")	
  
	Part.Diff <- tmp
  names(Part.Diff) <- paste("DIFF",names(Part.Diff),sep="_")	
	
	# -------------------------------------------------------------------------------------------------
	# [myEUI.raw] consists of the EUI of the TRB, the PRO, the EUI difference between [PRO} and [TRB]
	# -------------------------------------------------------------------------------------------------
	myEUI.raw     <- cbind(Part.Advn,Part.Base,Part.Diff)					# the EUI / EUI difference between the PRO and the TRB
	cat(paste("30: [",this.state,"]: get ",dim(myEUI.raw)[1]," by ",dim(myEUI.raw)[2]," [myEUI.raw]: through [myData.BASE] <--[myData.TRB] and [myData.ADVN] <-- [myData.PRO] and then [myEUI.raw] <- [myData.ADVN] - [myData.BASE] .\n",sep=""))
	cat(paste("30: [",this.state,"]: get ",dim(myEUI.raw)[1]," by ",dim(myEUI.raw)[2]," [myEUI.raw]: through [myData.BASE] <--[myData.TRB] and [myData.ADVN] <-- [myData.PRO] and then [myEUI.raw] <- [myData.ADVN] - [myData.BASE] .\n",sep=""),file=FL.LOG,append=TRUE)




	# -------------------------------------------------------------------------------------------------
	# assign the htgSys shares to each heating system type in this state: !!!!used the heating system shares!!!!
	# -------------------------------------------------------------------------------------------------
	idx.match.htgSys <- match(myEUI.raw[,"PRO_SS.HeatSys"],row.names(array.share.htgSys.state.revised))
	myEUI.raw[,"wgt.htgSys"] <- array.share.htgSys.state.revised[idx.match.htgSys,this.state]
	cat(paste("31: [",this.state,"]: attached htgsys shares to each models in [myEUI.raw].\n",sep=""))
	cat(paste("31: [",this.state,"]: attached htgsys shares to each models in [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)


	# -------------------------------------------------------------------------------------------------
	# assign the foundation shares to each heating system type in this state !!!!used the foundation shares!!!!
	# -------------------------------------------------------------------------------------------------
	idx.match.found <- match(myEUI.raw[,"PRO_SS.Foundation"],row.names(array.share.found.state.revised))
	myEUI.raw[,"wgt.found"] <- array.share.found.state.revised[idx.match.found,this.state]
	cat(paste("32: [",this.state,"]: attached foundation shares to each models in [myEUI.raw].\n",sep=""))
	cat(paste("32: [",this.state,"]: attached foundation shares to each models in [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)

	# -------------------------------------------------------------------------------------------------
	# calculate a cross product of the htgSys and foundation shares
	# -------------------------------------------------------------------------------------------------
	myEUI.raw[,"wgt.htgsys.found"] <- myEUI.raw[,"wgt.htgSys"] * myEUI.raw[,"wgt.found"]
	cat(paste("33: [",this.state,"]: computing a htgsys-foundation shares to each models in [myEUI.raw].\n",sep=""))
	cat(paste("33: [",this.state,"]: computing a htgsys-foundation shares to each models in [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)

	# -------------------------------------------------------------------------------------------------
	# add a field to label each level of each code item
	# -------------------------------------------------------------------------------------------------
	myEUI.raw[,"PRO_SS.PRO.CASE"] <- paste(myEUI.raw[,"PRO_SS.City"],myEUI.raw[,"PRO_SS.PRO.IDX"],myEUI.raw[,"PRO_SS.leakage_ratio"],sep="-")
	cat(paste("34: [",this.state,"]: add a CZ-CodeItem identification fields to each models in [myEUI.raw].\n",sep=""))
	cat(paste("34: [",this.state,"]: add a CZ-CodeItem identification fields to each models in [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)



	# -------------------------------------------------------------------------------------------------
	# OUTPUT EUI and deltaEUI of raw models beofe aggregation across foundation and climate zone
	#
	# save [myEUI.raw] to [FOA_EUI_dEUI_RawModel.csv] & [FOA_EUI_dEUI_RawModel.rds]
	# 
	# [myEUI.raw] has 24000 rows (1500 * 16)
	# -------------------------------------------------------------------------------------------------
	cat(paste("EUI DIFF of Raw Models before aggregating across htg sys and foundation,",sep=""),file=FL.EUI_dEUI.Raw.csv,append=TRUE)
	write.table(myEUI.raw,file=FL.EUI_dEUI.Raw.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	saveRDS(myEUI.raw,file=FL.EUI_dEUI.Raw.rds)
	cat(paste("35: [",this.state,"]: output [myEUI.raw].\n",sep=""))
	cat(paste("35: [",this.state,"]: output [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)

	#
	# [myEUI.raw.long] <- [myEUI.raw] for plotting the distribution of the code item purpose
	#
	field.varying    <- paste("PRO_",fields.num,sep="")	# only some of the PRO fields
	field.notVarying <- names(myEUI.raw)[!(names(myEUI.raw) %in% field.varying)]
	myEUI.raw.long <- melt(myEUI.raw,id.vars = field.notVarying,measure.vars = field.varying,variable.name = "Variable",value.name = "value")

		# 
		# Plotting the distributions of the code item values in the 24000 PRO 
		#
		# 1. ach50
		mySubset.4plot <- subset(myEUI.raw.long,subset=(Variable == "PRO_SS.leakage_ratio"))
		this.codeItem <- "leakage ratio"
		this.TRB      <- "TRB_SS.leakage_ratio"
		this.CZ       <- "PRO_SS.CZ"	
		names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
		p.plot01 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
		p.plot01 <- p.plot01 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
		p.plot01 <- p.plot01 + labs(x=this.codeItem,y="count",title=paste(paste("Sampled values of [",this.codeItem,"] in simulation at [",this.state,"]",sep=""),"\n",sep=""))
		p.plot01 <- p.plot01 + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 

		dev.set(2)
		plot(p.plot01)


	# -------------------------------------------------------------------------------------------------
	# AGGREGATION TRB EUI, PRO EUI, Delta EUI between TRB and PRO across foundation types and heating system types
	#
	# [myEUI.agg] <- [myEUI.raw] aggregating across foundation and heating systems
	#
	# [myEUI.agg] has 1500 rows where 1500 is  the number of randomnly drawn houses
	# NOTE: the fields name of [myEUI.agg] has been renamed here.  The file name used in plotting should refer here!!!!
	# -------------------------------------------------------------------------------------------------
	# weighted aggregation of the 5 measures across all htgsys and all foundations
	myEUI.agg.elec.TRB  <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"TRB_SS.tot.EUI.elec"], w = x[,"wgt.htgsys.found"]))
	myEUI.agg.gas.TRB   <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"TRB_SS.tot.EUI.gas"],  w = x[,"wgt.htgsys.found"]))
	myEUI.agg.tot.TRB   <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"TRB_SS.tot.EUI.sum"],  w = x[,"wgt.htgsys.found"]))
	
	myEUI.agg.elec.PRO  <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"PRO_SS.tot.EUI.elec"], w = x[,"wgt.htgsys.found"]))
	myEUI.agg.gas.PRO   <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"PRO_SS.tot.EUI.gas"],  w = x[,"wgt.htgsys.found"]))
	myEUI.agg.tot.PRO   <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"PRO_SS.tot.EUI.sum"],  w = x[,"wgt.htgsys.found"]))

	myEUI.agg.elec.DIFF <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"DIFF_SS.tot.EUI.elec"],w = x[,"wgt.htgsys.found"]))
	myEUI.agg.gas.DIFF  <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"DIFF_SS.tot.EUI.gas"], w = x[,"wgt.htgsys.found"]))
	myEUI.agg.tot.DIFF  <- sapply(split(myEUI.raw, myEUI.raw[,"PRO_SS.PRO.CASE"]), function(x) weighted.mean(x[,"DIFF_SS.tot.EUI.sum"], w = x[,"wgt.htgsys.found"]))
	
	      myEUI.agg  <- data.frame(cbind(myEUI.agg.elec.TRB,myEUI.agg.gas.TRB,myEUI.agg.tot.TRB,myEUI.agg.elec.PRO,myEUI.agg.gas.PRO,myEUI.agg.tot.PRO,myEUI.agg.elec.DIFF,myEUI.agg.gas.DIFF,myEUI.agg.tot.DIFF))
	names(myEUI.agg) <-                     c("EUI.elec.TRB",    "EUI.gas.TRB",    "EUI.tot.TRB",    "EUI.elec.PRO",    "EUI.gas.PRO",    "EUI.tot.PRO",    "EUI.elec.DIFF",    "EUI.gas.DIFF",    "EUI.tot.DIFF")
	cat(paste("35A: [",this.state,"]: ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] is derived from ",dim(myEUI.raw)[1]," by ",dim(myEUI.raw)[2]," [myEUI.raw]: aggregate the data across the htgsys and foundations for each CZ - codeitem.level combination for each models in [myEUI.raw].\n",sep=""))
	cat(paste("35A: [",this.state,"]: ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] is derived from ",dim(myEUI.raw)[1]," by ",dim(myEUI.raw)[2]," [myEUI.raw]: aggregate the data across the htgsys and foundations for each CZ - codeitem.level combination for each models in [myEUI.raw].\n",sep=""),file=FL.LOG,append=TRUE)




	# split the label field (which is the row names of [myEUI.agg]) into CZ, code name and code level
	myEUI.agg[,"SS.City"]    <- sub("(.*)-(.*)-(.*)","\\1",row.names(myEUI.agg),perl=TRUE)			
	myEUI.agg[,"SS.PRO.IDX"] <- sub("(.*)-(.*)-(.*)","\\2",row.names(myEUI.agg),perl=TRUE)			
	myEUI.agg[,"SS.PRO.NUM"] <- as.numeric(sub("PRO.duct","",sub("(.*)-(.*)-(.*)","\\2",row.names(myEUI.agg),perl=TRUE)))
	myEUI.agg[,"SS.STATE"]   <- this.state
	myEUI.agg[,"SS.CZ"]      <- myEPW.state[match(myEUI.agg[,"SS.City"],myEPW.state[,"CZ.city"]),"CZ.revised"]
	cat(paste("35B: [",this.state,"]: ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] is ready.\n",sep=""))
	cat(paste("35B: [",this.state,"]: ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] is ready.\n",sep=""),file=FL.LOG,append=TRUE)
	
	
	
	# creating a long format version of [myEUI.agg] for plotting purpose: re-arrange the 6 variables of [EUI.elec.PRO],[EUI.gas.PRO],[EUI.tot.PRO],[EUI.elec.DIFF],[EUI.gas.DIFF],[EUI.tot.DIFF] 
	id.vars        <- c("SS.City","SS.PRO.IDX","SS.PRO.NUM","SS.STATE","SS.CZ",grep("TRB",names(myEUI.agg),value=TRUE))	# include the baseline EUI columns and auxilary fields as the ID vars
	measure.vars   <- names(myEUI.agg)[-c(match(id.vars,names(myEUI.agg)))]							# the non-match subset of the fields: including the 6 PRO fields (elec, gas, tot EUI and DIFF)	
	myEUI.agg.long <- melt(myEUI.agg,id.vars = id.vars,measure.vars = measure.vars,variable.name = "Variable",value.name = "value")
	cat(paste("36: [",this.state,"]: have a ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] and a ",dim(myEUI.agg.long)[1]," by ",dim(myEUI.agg.long)[2]," [myEUI.agg.long] ready.\n",sep=""))
	cat(paste("36: [",this.state,"]: have a ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg] and a ",dim(myEUI.agg.long)[1]," by ",dim(myEUI.agg.long)[2]," [myEUI.agg.long] ready.\n",sep=""),file=FL.LOG,append=TRUE)

	#
	# add mean, median, CZ weighted TRB EUI to [myEUI.agg]
	#
	

	# ----------------------------------------------------------------------------------
	# save [myEUI.agg]
	# [myEUI.agg] has 1500 models
	# ----------------------------------------------------------------------------------
	cat(paste("EUI & Delta EUI,",sep=""),file=FL.EUI_dEUI.Agg.csv,append=TRUE)
	write.table(myEUI.agg,file=FL.EUI_dEUI.Agg.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	
	saveRDS(myEUI.agg,file=FL.EUI_dEUI.Agg.rds)
	cat(paste("37: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""))
	cat(paste("37: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""),file=FL.LOG,append=TRUE)

	


	# ----------------------------------------------------------------------------------
	# plotting series I
	# ----------------------------------------------------------------------------------
	count.plot <- 0
	for (label.CZ in sort(unique(myEUI.agg[,"SS.CZ"])))
	{
		count.plot <- count.plot + 1
		
		      mySubset.4plot  <- subset(myEUI.agg,subset=(SS.CZ == label.CZ),select = EUI.tot.PRO,drop=FALSE)
		      mySubset.4plot  <- mySubset.4plot[order(mySubset.4plot[,"EUI.tot.PRO"]),,drop=FALSE]
		      mySubset.4plot[,"Index"] <- seq(from=1,to=dim(mySubset.4plot)[1])
		names(mySubset.4plot) <- sub("EUI.tot.PRO","EUI",names(mySubset.4plot))
		baseline.EUI <- myEUI.agg[myEUI.agg[,"SS.CZ"] == label.CZ,"EUI.tot.TRB"][1]	# the corresponding baseline EUI
		baseline.IDX <- seq(1:dim(mySubset.4plot)[1])[mySubset.4plot[,"EUI"] > baseline.EUI][1]
		
		# write the ranked models out
		cat(paste("(",label.CZ,") (",this.state,")",sep=""),file=FL.EUI_dEUI.rank.csv,append=TRUE)
		write.table(mySubset.4plot,file=FL.EUI_dEUI.rank.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
		cat(paste("\n\n",sep=""),file=FL.EUI_dEUI.rank.csv,append=TRUE)
		
		plot.obj <- xyplot(EUI~Index,data=mySubset.4plot,
				   horizontal = FALSE,
				   type       = "h",
				   layout     = c(0,1),
				   as.table   = TRUE,
				   between    = list(y=0.5),
				   xlab       = "Index of Models",
				   ylab       = "EUI (kBtu/sf)",
				   main       = paste("(",this.state,"): IECC Regulated end-use EUI (kBtu/sf) of ",dim(mySubset.4plot)[1]," pseudo houses sampled at (",label.CZ,")\nHorizontal line indicates the EUI for a 2009 IECC prescriptive code-compliant prototype",sep=""),
				   cex        = 0.5,
				   pch        = 19,
				   col        = "blue",

			   panel = function(x,y,...) {
				   panel.xyplot(x, y,...)
				   panel.abline(h=baseline.EUI,col=c("red"))
				   panel.abline(v=baseline.IDX,col=c("red"))
				   panel.text(0,baseline.EUI,adj=c(-0.25,-0.25),labels=paste("2009 IECC EUI of (",label.CZ,")=",round(baseline.EUI,digits=2)," (kBtu/sf)",sep=""),col="red")}	
				   )
				   
				   # plot(plot.obj)

			   	
		# assign a plot number of the plot
		command.string <- paste(paste("p.plot00",count.plot,sep="")," <- plot.obj",sep="") 
		eval(parse(text=command.string))		
		
		# plotting
		dev.set(3)
		plot(plot.obj)
	}
	cat(paste("39: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""))
	cat(paste("39: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""),file=FL.LOG,append=TRUE)
	


	# -------------------------------------------------------------------------------------------------
	# -------------------------------------------------------------------------------------------------
	# -------------------------------------------------------------------------------------------------
	# one plot for a state without distinguish the CZs.  another plot with all model EUI plotted and a weighted baseline EUI plotted
	# -------------------------------------------------------------------------------------------------
	# -------------------------------------------------------------------------------------------------
	# -------------------------------------------------------------------------------------------------
	# only keep the non-zero CZ share in the state
	thisState.CZ.shares <- array.share.CZ.state.revised[array.share.CZ.state.revised[,this.state,drop=FALSE]>0,this.state,drop=FALSE]
	
	# in [myEUI.agg], there is a field "EUI.tot.TRB" which has only 2 or 3 unique numbers, one for each CZ, retrieve the frst row to reach
	thisState.baselineEUI  <- myEUI.agg[diff(c(0,myEUI.agg[,"EUI.tot.TRB"])) != 0,c("SS.CZ","EUI.tot.TRB"),drop=FALSE]
	thisState.baselineEUI[,"CZ.share"] <- thisState.CZ.shares[thisState.baselineEUI[,"SS.CZ"],this.state,drop=FALSE]
	thisState.weightedEUI <- weighted.mean(thisState.baselineEUI[,"EUI.tot.TRB"], w = thisState.baselineEUI[,"CZ.share"])
	cat(paste("39A: [",this.state,"]: baseline EUI in CZs [thisState.baselineEUI] and weighted baselineEUI of the state in [thisState.weightedEUI].\n",sep=""))
	cat(paste("39A: [",this.state,"]: baseline EUI in CZs [thisState.baselineEUI] and weighted baselineEUI of the state in [thisState.weightedEUI].\n",sep=""),file=FL.LOG,append=TRUE)

	
	
	# plot the ranked EUI of the state without distinguishing the CZs
	      mySubset.4plot  <- myEUI.agg[,c("SS.CZ","EUI.tot.PRO"),drop=FALSE]
	      mySubset.4plot  <- mySubset.4plot[order(mySubset.4plot[,"EUI.tot.PRO"]),,drop=FALSE]
	      mySubset.4plot[,"Index"] <- seq(from=1,to=dim(mySubset.4plot)[1])
	names(mySubset.4plot) <- sub("EUI.tot.PRO","EUI",names(mySubset.4plot))	# note field name changed from "EUI.tot.PRO" to "EUI"

	# -------------------------------------------------------------------------------------------------
	# add the CZ fraction weighted TRB EUI for the state
	# -------------------------------------------------------------------------------------------------
	mySubset.4plot[,"WgtTRBEUI"] <- thisState.weightedEUI
	
	# -------------------------------------------------------------------------------------------------
	# add the TRB EUI for each of the CZ present in the state
	# -------------------------------------------------------------------------------------------------	
	for (this.CZ in unique(mySubset.4plot[,"SS.CZ"]))
	{
		mySubset.4plot[,paste(this.CZ,".TRBEUI",sep="")] <- thisState.baselineEUI[thisState.baselineEUI[,"SS.CZ"] == this.CZ,"EUI.tot.TRB"]
	}
	cat(paste("39B: [",this.state,"]: added baseline EUI to [mySubset.4plot].\n",sep=""))
	cat(paste("39B: [",this.state,"]: added baseline EUI to [mySubset.4plot].\n",sep=""),file=FL.LOG,append=TRUE)

	# -------------------------------------------------------------------------------------------------
	# calculate the mean and median of the 1500 EUIs
	# -------------------------------------------------------------------------------------------------
	thisState.meanEUI   <- mean(mySubset.4plot[,"EUI"],na.rm=TRUE)
	thisState.medianEUI <- median(mySubset.4plot[,"EUI"],na.rm=TRUE)
	mySubset.4plot[,"meanEUI"]   <- thisState.meanEUI
	mySubset.4plot[,"medianEUI"] <- thisState.medianEUI	
	cat(paste("39C: [",this.state,"]: calculate the mean EUI of ",dim(myEUI.agg)[1]," randomly sampled houses.\n",sep=""))
	cat(paste("39C: [",this.state,"]: calculate the mean EUI of ",dim(myEUI.agg)[1]," randomly sampled houses.\n",sep=""),file=FL.LOG,append=TRUE)


	# write the ranked models out
	cat(paste("(",this.state,"),",sep=""),file=FL.EUI_dEUI.rank.csv,append=TRUE)
	write.table(mySubset.4plot,file=FL.EUI_dEUI.rank.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	cat(paste("\n\n",sep=""),file=FL.EUI_dEUI.rank.csv,append=TRUE)
	cat(paste("39D: [",this.state,"]: ranked EUI are written out.\n",sep=""))
	cat(paste("39D: [",this.state,"]: ranked EUI are written out.\n",sep=""),file=FL.LOG,append=TRUE)


	# find the index in the ranked EUI array for  mean, median and CZ weighted EUI
	thisIndex.meanEUI     <- seq(1:dim(mySubset.4plot)[1])[mySubset.4plot[,"EUI"] > thisState.meanEUI][1]
	thisIndex.medianEUI   <- seq(1:dim(mySubset.4plot)[1])[mySubset.4plot[,"EUI"] > thisState.medianEUI][1]
	thisIndex.weightedEUI <- seq(1:dim(mySubset.4plot)[1])[mySubset.4plot[,"EUI"] > thisState.weightedEUI][1]
	cat(paste("39E: [",this.state,"]: the index in the ranked EUI df of the mean and median EUI and the CZ weighted TRB EUI.\n",sep=""))
	cat(paste("39E: [",this.state,"]: the index in the ranked EUI df of the mean and median EUI and the CZ weighted TRB EUI.\n",sep=""),file=FL.LOG,append=TRUE)
	
	
	plot.obj <- xyplot(EUI~Index,data=mySubset.4plot,
			   horizontal = FALSE,
			   type       = "h",
			   layout     = c(0,1),
			   as.table   = TRUE,
			   between    = list(y=0.5),
			   xlab       = "Index of Models",
			   ylab       = "EUI (kBtu/sf)",
			   main       = paste("(",this.state,"): IECC Regulated end-use EUI (kBtu/sf) of ",dim(mySubset.4plot)[1]," pseudo houses sampled\nRed line indicates the EUI of code complianced house\nGreen dotted line indicate the mean of the simulated houses\nCyan dotted line indicate the median of the simulated houses",sep=""),
			 # main       = paste("(",label.CZ,")-(",this.state,"): IECC Regulated end-use EUI of the models based on randow sampling of code items field observations",sep=""),
			   cex        = 0.5,
			   pch        = 19,
			   col        = "blue",

		   panel = function(x,y,...) {
			   panel.xyplot(x, y,...)
			   panel.abline(h=thisState.weightedEUI,col=c("red"),  lty=1)
			   panel.abline(h=thisState.meanEUI,    col=c("green"),lty=2)
			   panel.abline(h=thisState.medianEUI,  col=c("cyan"), lty=2)
			   
			   panel.abline(v=thisIndex.weightedEUI,col=c("red"),  lty=1)
			   panel.abline(v=thisIndex.meanEUI,    col=c("green"),lty=2)
			   panel.abline(v=thisIndex.medianEUI,  col=c("cyan"), lty=2)			   
			   panel.text(0,thisState.weightedEUI,adj=c(-0.25,-0.25),labels=paste("CZ Weighted Code EUI = ",       round(thisState.weightedEUI,digits=2)," (kBtu/sf)",sep=""),col="red")		
			   panel.text(0,thisState.meanEUI,    adj=c(-0.25,-0.25),labels=paste("Mean of the simulated EUI = ",  round(thisState.meanEUI,    digits=2)," (kBtu/sf)",sep=""),col="green")
			   panel.text(0,thisState.medianEUI,  adj=c(-0.25,-0.25),labels=paste("median of the simulated EUI = ",round(thisState.medianEUI,  digits=2)," (kBtu/sf)",sep=""),col="cyan")}
			   )


		# assign a plot number of the plot
		command.string <- paste(paste("p.plot00",count.plot,sep="")," <- plot.obj",sep="") 
		eval(parse(text=command.string))		

		# plotting
		dev.set(3)
		plot(plot.obj)	
	cat(paste("39E: [",this.state,"]: ranked EUI are plotted in terms of the CZs in the state.\n",sep=""))
	cat(paste("39E: [",this.state,"]: ranked EUI are plotted in terms of the CZs in the state.\n",sep=""),file=FL.LOG,append=TRUE)




	# ----------------------------------------------------------------------------------
	# plotting series II
	# ----------------------------------------------------------------------------------
	# 11. IECC Regulated end-use Gas EUI
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.gas.PRO"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Gas EUI"
	this.TRB      <- "EUI.gas.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot31 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot31 <- p.plot31 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot31 <- p.plot31 + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem," at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot31 <- p.plot31 + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	p.plot31 <- p.plot31 + geom_text(aes(x=CodeCompliance,y=y,label = round(CodeCompliance,digits=2)),data = mySubset.4plot,colour="red",size=5)	

	# 12. IECC Regulated end-use Elec EUI
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.elec.PRO"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Elec EUI"
	this.TRB      <- "EUI.elec.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot32 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot32 <- p.plot32 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot32 <- p.plot32 + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem," at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot32 <- p.plot32 + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	p.plot32 <- p.plot32 + geom_text(aes(x=CodeCompliance,y=y,label = round(CodeCompliance,digits=2)),data = mySubset.4plot,colour="red",size=5)	
	
	
	# 13. IECC Regulated end-use Total EUI
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.PRO"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.TRB      <- "EUI.tot.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot33 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot33 <- p.plot33 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot33 <- p.plot33 + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem," at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot33 <- p.plot33 + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	p.plot33 <- p.plot33 + geom_text(aes(x=CodeCompliance,y=y,label = round(CodeCompliance,digits=2)),data = mySubset.4plot,colour="red",size=5)	

	# 14. IECC Regulated end-use Total EUI
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.PRO"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.TRB      <- "EUI.tot.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot34 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram()
	p.plot34 <- p.plot34 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot34 <- p.plot34 + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem," at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot34 <- p.plot34 + geom_vline(xintercept=thisState.weightedEUI,colour="black") 		
	p.plot34 <- p.plot34 + annotate("text",label=round(thisState.weightedEUI,digits=2),x=thisState.weightedEUI,y=0,colour="black")	
      # p.plot34 <- p.plot34 + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 		
      # p.plot34 <- p.plot34 + geom_text(aes(x=CodeCompliance,y=y,label = round(CodeCompliance,digits=2)),data = mySubset.4plot,colour="red",size=5)	
	

	# 15. IECC Regulated end-use Total EUI
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.PRO"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.TRB      <- "EUI.tot.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot35 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue)) + geom_histogram(colour="black",fill="black")
	p.plot35 <- p.plot35 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot35 <- p.plot35 + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem," at [",this.state,"]",sep=""),"\nblack line: code complaince EUI\ngreen line: average EUI of the simulated houses",sep=""))
	p.plot35 <- p.plot35 + geom_vline(aes(xintercept=thisState.weightedEUI), , colour="red", data =mySubset.4plot) 
	p.plot35 <- p.plot35 + geom_vline(xintercept=thisState.meanEUI,    colour="green") 
	p.plot35 <- p.plot35 + geom_vline(xintercept=thisState.medianEUI,  colour="cyan") 	
	p.plot35 <- p.plot35 + annotate("text",label=round(thisState.weightedEUI,digits=2),x=thisState.weightedEUI,y=0,colour="red")
      # p.plot35 <- p.plot35 + geom_text(aes(x=CodeCompliance,y=y,label = round(CodeCompliance,digits=2)),data = mySubset.4plot,colour="red",size=5)	






	# 16. Delta IECC Regulated end-use Gas EUI between TRB and PRO
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.gas.DIFF"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Gas EUI"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot36 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot36 <- p.plot36 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot36 <- p.plot36 + labs(x=this.codeItem,y="count",title=paste(paste("Difference of [",this.codeItem,"] from the Baseline in simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot36 <- p.plot36 + geom_vline(aes(xintercept=0), data =mySubset.4plot) 
	

	# 17. Delta IECC Regulated end-use Elec EUI between TRB and PRO
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.elec.DIFF"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Elec EUI"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot37 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot37 <- p.plot37 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot37 <- p.plot37 + labs(x=this.codeItem,y="count",title=paste(paste("Difference of [",this.codeItem,"] from the Baseline in simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot37 <- p.plot37 + geom_vline(aes(xintercept=0), data =mySubset.4plot) 
	
	
	# 18. Delta IECC Regulated end-use Total EUI between TRB and PRO
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.DIFF"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot38 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram() + facet_wrap(~ CZ,ncol = 1)
	p.plot38 <- p.plot38 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot38 <- p.plot38 + labs(x=this.codeItem,y="count",title=paste(paste("Difference of [",this.codeItem,"] from the Baseline in simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot38 <- p.plot38 + geom_vline(aes(xintercept=0), data =mySubset.4plot) 
	
	# 19. Delta IECC Regulated end-use Total EUI between TRB and PRO
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.DIFF"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot39 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue,colour=factor(CZ),fill=factor(CZ))) + geom_histogram()
	p.plot39 <- p.plot39 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot39 <- p.plot39 + labs(x=this.codeItem,y="count",title=paste(paste("Difference of [",this.codeItem,"] from the Baseline in simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot39 <- p.plot39 + geom_vline(aes(xintercept=0), data =mySubset.4plot) 
	
	# 20. Delta IECC Regulated end-use Total EUI between TRB and PRO
	mySubset.4plot <- subset(myEUI.agg.long,subset=(Variable == "EUI.tot.DIFF"))
	mySubset.4plot[,"y"] <- 0
	this.codeItem <- "Aggregated IECC Regulated end-use Total EUI"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",sub("value","KeyCodeValue",names(mySubset.4plot))))
	p.plot40 <- ggplot(data=mySubset.4plot,aes(x=KeyCodeValue)) + geom_histogram(colour="black",fill="black")
	p.plot40 <- p.plot40 + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="")
	p.plot40 <- p.plot40 + labs(x=this.codeItem,y="count",title=paste(paste("Difference of [",this.codeItem,"] from the Baseline in simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot40 <- p.plot40 + geom_vline(aes(xintercept=0), colour="red",data =mySubset.4plot) 
	p.plot40 <- p.plot40 + annotate("text",label=0,x=0,y=0,colour="red")
	
	
	
	dev.set(3)
	multiplot(p.plot31,p.plot32,p.plot33,p.plot36,p.plot37,p.plot38,cols=2)
	multiplot(p.plot34,p.plot35,p.plot39,p.plot40,cols=2)
	cat(paste("40: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""))
	cat(paste("40: [",this.state,"]: output ",dim(myEUI.agg)[1]," by ",dim(myEUI.agg)[2]," [myEUI.agg].\n",sep=""),file=FL.LOG,append=TRUE)




	
	# elec EUI
	this.var    <- "EUI.elec.PRO"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))				
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot1A <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1A <- p.plot1A + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1A <- p.plot1A + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))	
	p.plot1A <- p.plot1A + geom_vline(aes(xintercept=EUI.elec.TRB), data = mySubset)

	p.plot1B <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1B <- p.plot1B + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1B <- p.plot1B + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))	
	p.plot1B <- p.plot1B + geom_vline(aes(xintercept=EUI.elec.TRB), data = mySubset)

	# gas EUI
	this.var <- "EUI.gas.PRO"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot1C <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1C <- p.plot1C + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1C <- p.plot1C + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			
	p.plot1C <- p.plot1C + geom_vline(aes(xintercept=EUI.gas.TRB), data = mySubset)

	p.plot1D <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1D <- p.plot1D + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1D <- p.plot1D + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))	
	p.plot1D <- p.plot1D + geom_vline(aes(xintercept=EUI.gas.TRB), data = mySubset)


	# sum EUI
	this.var <- "EUI.tot.PRO"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot1E <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1E <- p.plot1E + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1E <- p.plot1E + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))	
	p.plot1E <- p.plot1E + geom_vline(aes(xintercept=EUI.tot.TRB), data = mySubset)

	p.plot1F <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot1F <- p.plot1F + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot1F <- p.plot1F + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))	
	p.plot1F <- p.plot1F + geom_vline(aes(xintercept=EUI.tot.TRB), data = mySubset)

	dev.set(3)
		multiplot(p.plot1A,p.plot1C,p.plot1E,p.plot1B,p.plot1D,p.plot1F,cols=2)
	cat(paste("41: [",this.state,"]: plot the EUI distribution.\n",sep=""))
	cat(paste("41: [",this.state,"]: plot the EUI distribution.\n",sep=""),file=FL.LOG,append=TRUE)


	# elec delta EUI
	this.var <- "EUI.elec.DIFF"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))				
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot2A <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2A <- p.plot2A + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2A <- p.plot2A + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			

	p.plot2B <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2B <- p.plot2B + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2B <- p.plot2B + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			

	# gas delta EUI
	this.var <- "EUI.gas.DIFF"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot2C <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2C <- p.plot2C + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2C <- p.plot2C + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			

	p.plot2D <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2D <- p.plot2D + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2D <- p.plot2D + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			


	# sum delta EUI
	this.var <- "EUI.tot.DIFF"
	this.string <- sub("DIFF","DELTA",sub("(.*)\\.(.*)\\.(.*)","(\\3)-(\\2)-(\\1)",this.var,perl=TRUE))
	mySubset <- subset(myEUI.agg.long,subset=(Variable==this.var))
	p.plot2E <- qplot(data=mySubset,value,facets=SS.STATE~SS.CZ,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2E <- p.plot2E + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2E <- p.plot2E + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			

	p.plot2F <- qplot(data=mySubset,value,facets=SS.STATE~.,colour=factor(SS.CZ),fill=factor(SS.CZ),geom="histogram") 
	p.plot2F <- p.plot2F + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right") 
	p.plot2F <- p.plot2F + labs(x=this.string,y="Value (kBtu/sf)",title=paste("Distribution of ",this.string,"\n",sep=""))			

	dev.set(3)
		multiplot(p.plot2A,p.plot2C,p.plot2E,p.plot2B,p.plot2D,p.plot2F,cols=2)
	cat(paste("42: [",this.state,"]: plot the Delta EUI distribution.\n",sep=""))
	cat(paste("42: [",this.state,"]: plot the Delta EUI distribution.\n",sep=""),file=FL.LOG,append=TRUE)



	#
	# more plot: step function etc
	#
	# p.plot3A <- xyplot(value~index,data=mySubset,type="h")



	# ---------------------------------------------------------------------------------
	# make some statistics on PRO data across all code item
	# ---------------------------------------------------------------------------------
	# all htgsys and all foundation
	mySTAT.CZ.EUI.elec <- as.data.frame(tapply(myEUI.agg[,"EUI.elec.PRO"],list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	mySTAT.CZ.EUI.gas  <- as.data.frame(tapply(myEUI.agg[,"EUI.gas.PRO"], list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	mySTAT.CZ.EUI.sum  <- as.data.frame(tapply(myEUI.agg[,"EUI.tot.PRO"], list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	names(mySTAT.CZ.EUI.elec) <- "average EUI"
	names(mySTAT.CZ.EUI.gas ) <- "average EUI"
	names(mySTAT.CZ.EUI.sum ) <- "average EUI"

	mySTAT.CZ.deltaEUI.elec <- as.data.frame(tapply(myEUI.agg[,"EUI.elec.DIFF"],list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	mySTAT.CZ.deltaEUI.gas  <- as.data.frame(tapply(myEUI.agg[,"EUI.gas.DIFF"], list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	mySTAT.CZ.deltaEUI.sum  <- as.data.frame(tapply(myEUI.agg[,"EUI.tot.DIFF"], list(myEUI.agg[,"SS.CZ"]),mean,na.rm=TRUE))
	names(mySTAT.CZ.deltaEUI.elec) <- "average Delta EUI"
	names(mySTAT.CZ.deltaEUI.gas ) <- "average Delta EUI"
	names(mySTAT.CZ.deltaEUI.sum ) <- "average Delta EUI"
	cat(paste("43: [",this.state,"]: statistics.\n",sep=""))
	cat(paste("43: [",this.state,"]: statistics.\n",sep=""),file=FL.LOG,append=TRUE)
	
	



	#
	# calculate the EUI changes of the models from the baseline due to duct leakage
	#
	myEUI.agg[,"EUI.elec.PntChange"] <- 100 * (myEUI.agg[,"EUI.elec.TRB"] - myEUI.agg[,"EUI.elec.PRO"]) / myEUI.agg[,"EUI.elec.TRB"]
	myEUI.agg[,"EUI.gas.PntChange"]  <- 100 * (myEUI.agg[,"EUI.gas.TRB"]  - myEUI.agg[,"EUI.gas.PRO"])  / myEUI.agg[,"EUI.gas.TRB"]
	myEUI.agg[,"EUI.tot.PntChange"]  <- 100 * (myEUI.agg[,"EUI.tot.TRB"]  - myEUI.agg[,"EUI.tot.PRO"])  / myEUI.agg[,"EUI.tot.TRB"]
	cat(paste("44: [",this.state,"]: calculate the EUI change percentage.\n",sep=""))
	cat(paste("44: [",this.state,"]: calculate the EUI change percentage.\n",sep=""),file=FL.LOG,append=TRUE)
	
	myEUI.agg[,"ID"]            <- row.names(myEUI.agg)
	myEUI.agg[,"City"]          <- sub("(.*)-(.*)-(.*)","\\1",myEUI.agg[,"ID"])
	myEUI.agg[,"INDEX"]         <- sub("(.*)-(.*)-(.*)","\\2",myEUI.agg[,"ID"])
	myEUI.agg[,"leakage.ratio"] <- sub("(.*)-(.*)-(.*)","\\3",myEUI.agg[,"ID"])
	cat(paste("45: [",this.state,"]: break down the ID field.\n",sep=""))
	cat(paste("45: [",this.state,"]: break down the ID field.\n",sep=""),file=FL.LOG,append=TRUE)
	
	cat(paste("duct leakage,",sep=""),file=FL.duct_EUI_change.csv,append=TRUE)
	write.table(myEUI.agg,file=FL.duct_EUI_change.csv,sep=",",row.names=TRUE,col.names=TRUE,append=TRUE)
	saveRDS(myEUI.agg,file=FL.duct_EUI_change.rds)
	cat(paste("44: [",this.state,"]: duct leakage EUI change.\n",sep=""))
	cat(paste("44: [",this.state,"]: duct leakage EUI change.\n",sep=""),file=FL.LOG,append=TRUE)

	#
	# turn wide to long format for plotting the duct sealing EUI for checking purpose
	#
	field.varying      <- c("EUI.elec.PRO","EUI.gas.PRO","EUI.tot.PRO","EUI.elec.DIFF","EUI.gas.DIFF","EUI.tot.DIFF","EUI.elec.PntChange","EUI.gas.PntChange","EUI.tot.PntChange")
	field.notVarying   <- names(myEUI.agg)[!(names(myEUI.agg) %in% field.varying)]
	myEUI.new.agg.long <- melt(myEUI.agg,id.vars = field.notVarying,measure.vars = field.varying,variable.name = "Variable",value.name = "value")
	if(is.character(myEUI.new.agg.long[,"leakage.ratio"])){myEUI.new.agg.long[,"leakage.ratio"] <- as.numeric(myEUI.new.agg.long[,"leakage.ratio"])}

	#
	# 1. plotting electricity EUI
	#
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.elec.PRO"))
	this.codeItem <- "Electricity EUI"
	this.TRB      <- "EUI.elec.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",names(mySubset.4plot)))
	# EUI
	p.plot.elec.A <- ggplot(data=mySubset.4plot,aes(x=value,colour=CZ,fill=CZ)) + geom_histogram() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.elec.A <- p.plot.elec.A + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.elec.A <- p.plot.elec.A + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem,"] of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.elec.A <- p.plot.elec.A + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	
	# EUI vs leakage ratio
	p.plot.elec.B <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.elec.B <- p.plot.elec.B + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.elec.B <- p.plot.elec.B + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.elec.B <- p.plot.elec.B + geom_hline(aes(yintercept=CodeCompliance), data =mySubset.4plot) 
	
	# EUI Change % vs leakage ratio
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.elec.PntChange"))
	this.codeItem <- "Electricity EUI Change %"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",names(mySubset.4plot))	
	p.plot.elec.C <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.elec.C <- p.plot.elec.C + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.elec.C <- p.plot.elec.C + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.elec.C <- p.plot.elec.C + geom_hline(aes(yintercept=0), data =mySubset.4plot) 



	#
	# 2. plotting gas EUI
	#
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.gas.PRO"))
	this.codeItem <- "Gas EUI"
	this.TRB      <- "EUI.gas.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",names(mySubset.4plot)))
	# EUI
	p.plot.gas.A <- ggplot(data=mySubset.4plot,aes(x=value,colour=CZ,fill=CZ)) + geom_histogram() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.gas.A <- p.plot.gas.A + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.gas.A <- p.plot.gas.A + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem,"] of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.gas.A <- p.plot.gas.A + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	
	# EUI vs leakage ratio
	p.plot.gas.B <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.gas.B <- p.plot.gas.B + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.gas.B <- p.plot.gas.B + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.gas.B <- p.plot.gas.B + geom_hline(aes(yintercept=CodeCompliance), data =mySubset.4plot) 

	# EUI Change % vs leakage ratio
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.gas.PntChange"))
	this.codeItem <- "Gas EUI Change %"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",names(mySubset.4plot))	
	p.plot.gas.C <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.gas.C <- p.plot.gas.C + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.gas.C <- p.plot.gas.C + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.gas.C <- p.plot.gas.C + geom_hline(aes(yintercept=0), data =mySubset.4plot) 



	#
	# 3. plotting total EUI
	#
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.tot.PRO"))
	this.codeItem <- "Total EUI"
	this.TRB      <- "EUI.tot.TRB"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",sub(this.TRB,"CodeCompliance",names(mySubset.4plot)))
	# EUI
	p.plot.tot.A <- ggplot(data=mySubset.4plot,aes(x=value,colour=CZ,fill=CZ)) + geom_histogram() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.tot.A <- p.plot.tot.A + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.tot.A <- p.plot.tot.A + labs(x=this.codeItem,y="count",title=paste(paste(this.codeItem,"] of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.tot.A <- p.plot.tot.A + geom_vline(aes(xintercept=CodeCompliance), data =mySubset.4plot) 
	
	# EUI vs leakage ratio
	p.plot.tot.B <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.tot.B <- p.plot.tot.B + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.tot.B <- p.plot.tot.B + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.tot.B <- p.plot.tot.B + geom_hline(aes(yintercept=CodeCompliance), data =mySubset.4plot) 

	# EUI Change % vs leakage ratio
	mySubset.4plot <- subset(myEUI.new.agg.long,subset=(Variable == "EUI.tot.PntChange"))
	this.codeItem <- "Total EUI Change %"
	this.CZ       <- "SS.CZ"	
	names(mySubset.4plot) <- sub(this.CZ,"CZ",names(mySubset.4plot))	
	p.plot.tot.C <- ggplot(data=mySubset.4plot,aes(x=leakage.ratio,y=value,colour=CZ,fill=CZ)) + geom_point() # + facet_wrap(~ CZ,ncol = 1)
	p.plot.tot.C <- p.plot.tot.C + theme(axis.text.x = element_text(angle=90,color="black",size=10),axis.text.y = element_text(color="black"),legend.position="right")
	p.plot.tot.C <- p.plot.tot.C + labs(x="Duct Leakage Ratio",y=this.codeItem,title=paste(paste(this.codeItem," of duct leakage simulation at [",this.state,"]",sep=""),"\n",sep=""))
	p.plot.tot.C <- p.plot.tot.C + geom_hline(aes(yintercept=0), data =mySubset.4plot) 
	
	dev.set(4)
		multiplot(p.plot.elec.A,p.plot.gas.A,p.plot.tot.A,p.plot.elec.B,p.plot.gas.B,p.plot.tot.B,p.plot.elec.C,p.plot.gas.C,p.plot.tot.C,cols=3)


	 
	dev.off(2)
	dev.off(3)
	dev.off(4)
	# ---------------------------------------------------------------------------------
	# save the data in R object
	# ---------------------------------------------------------------------------------
	save(list = ls(all=TRUE),file=FL.Rdata)


}	# end of state loop


# -------------------------------------------------------------------------------------------------
# time used for completing this script
# -------------------------------------------------------------------------------------------------
End.time  <- Sys.time()
Diff.time <- End.time - Start.time
Diff.time

cat(paste("\n19. 06A_duct_data_analysis.R is finished successfully at ",End.time,"!\n",sep=" "))
cat(paste("\n19. 06A_duct_data_analysis.R is finished successfully at ",End.time,"!\n",sep=" "),file=FL.LOG,append=TRUE)

cat(paste("\n20. Processing time for [06A_duct_data_analysis.R] is ",as.numeric(Diff.time, units="mins")," minutes\n",sep=" "))
cat(paste("\n20. Processing time for [06A_duct_data_analysis.R] is ",as.numeric(Diff.time, units="mins")," minutes\n",sep=" "),file=FL.LOG,append=TRUE)

#
# put run related information into the log file
#
cat(paste("21. This run was conducted in ",.Platform$OS.type,"\n",sep=""));
cat(paste("21. This run was conducted in ",.Platform$OS.type,"\n",sep=""),file=FL.LOG,append=TRUE);

# get the version of R used for this computation and the latest version released
current.Rversion <- R.Version()$version.string
tmp = readLines("http://cran.r-project.org/sources.html")
rls = tmp[grep("latest release", tmp) + 1L]			# the version number is in the next line of 'The latest release'
latest.Rversion  <- gsub("(.*R-|\\.tar\\.gz.*)", "", rls)	# "The latest release: R-2.13.0.tar.gz"
if (latest.Rversion != current.Rversion)
{
	cat(paste("\n\n22. you may want to upgrade R from the version you are using [",current.Rversion,"] to the latest version of [",latest.Rversion,"]\n",sep=""));
	cat(paste("\n\n22. you may want to upgrade R from the version you are using [",current.Rversion,"] to the latest version of [",latest.Rversion,"]\n",sep=""),file=FL.LOG,append=TRUE);
}else{
	cat(paste("\n\n22. The R version you are using is the latest version released so far!\n",sep=""))
	cat(paste("\n\n22. The R version you are using is the latest version released so far!\n",sep=""),file=FL.LOG,append=TRUE)
}



# get the version information of the attached libraries
cat(paste("\n\nThe information of the packages you used for this calculation:\n"))
cat(paste("\n\nThe information of the packages you used for this calculation:\n"),file=FL.LOG,append=TRUE)
tmp <- sessionInfo()
pkg.loaded <- tmp$otherPkgs
no.pkg.loaded <- length(pkg.loaded)
for (i in seq(1,no.pkg.loaded))
{
	cat(paste(pkg.loaded[[i]]$Package,":",pkg.loaded[[i]]$Version," ",pkg.loaded[[i]]$Date,"\n",sep=" "))
	cat(paste(pkg.loaded[[i]]$Package,":",pkg.loaded[[i]]$Version," ",pkg.loaded[[i]]$Date,"\n",sep=" "),file=FL.LOG,append=TRUE)
}






#
# old strategy
#
# ----------------------------------------------------------------------------------------------------------------------------------------------------------
# August 22, 2014: the following sampling design for detecting IECC Regulated end-use energy change is replaced by a new design to detect code item difference.
# ----------------------------------------------------------------------------------------------------------------------------------------------------------
# # randomly draw 1000 combinations
# sample.ceiling     <- sample(r_ceiling,  no.sample,replace = TRUE, prob = p_ceiling)
# sample.wall        <- sample(r_wall,     no.sample,replace = TRUE, prob = p_wall)
# sample.floor       <- sample(r_floor,    no.sample,replace = TRUE, prob = p_floor)
# sample.u_window    <- sample(u_window,   no.sample,replace = TRUE, prob = p_u_window)
# sample.shgc_window <- sample(shgc_window,no.sample,replace = TRUE, prob = p_shgc_window)
# sample.f_inc_hw    <- sample(f_inc_hw,   no.sample,replace = TRUE, prob = p_f_inc_hw)
# sample.ach50       <- sample(ach50,      no.sample,replace = TRUE, prob = p_ach50)
# sample.bsmtwall    <- sample(r_bsmtwall, no.sample,replace = TRUE, prob = p_bsmtwall)
# sample.slab        <- sample(r_slab,     no.sample,replace = TRUE, prob = p_slab)
# cat(paste("5. ",no.sample," random samples are drawn for  CZ ",this.city,"(",this.CZ,") are defined!\n",sep=""))
# cat(paste("5. ",no.sample," random samples are drawn for  CZ ",this.city,"(",this.CZ,") are defined!\n",sep=""),file=FL.LOG,append=TRUE)
# 
# # parm data frame
# if(no.CZ == 1)
# {
# 	df.parm <- data.frame(CZ_city       = rep(this.city,   no.sample),
# 			      CZ_label      = rep(this.CZ,     no.sample),
# 			      CZ_moist      = rep(this.moist,  no.sample),
# 			      CZ_weather    = rep(this.weather,no.sample),
# 			      tag           = rep(htgSys,      each = no.sample/length(htgSys)),			# assign equal number of cases to each of the four heating systems
# 			      foundation    = rep(foundations, each = no.sample/(length(foundations)*length(htgSys))),	# assign equal number of cases to each of the four heating systems
# 	                      r_ceiling     = sample.ceiling,
# 	                      r_wall        = sample.wall,
# 	                      r_floor       = sample.floor,
# 	                      u_window      = sample.u_window,
# 	                      shgc_window   = sample.shgc_window,
# 	                      f_inc_hw      = sample.f_inc_hw,
# 	                      ach50         = sample.ach50,
# 	                      r_bsmtwall    = sample.bsmtwall,
# 	                      r_slab        = sample.slab)
# }else{
# 	df.parm <- rbind(df.parm,
# 	           data.frame(CZ_city       = rep(this.city,   no.sample),
# 			      CZ_label      = rep(this.CZ,     no.sample),
# 			      CZ_moist      = rep(this.moist,  no.sample),
# 			      CZ_weather    = rep(this.weather,no.sample),
# 			      tag           = rep(htgSys,      each = no.sample/length(htgSys)),			# assign equal number of cases to each of the four heating systems
# 			      foundation    = rep(foundations, each = no.sample/(length(foundations)*length(htgSys))),	# assign equal number of cases to each of the four heating systems
# 	                      r_ceiling     = sample.ceiling,
# 	                      r_wall        = sample.wall,
# 	                      r_floor       = sample.floor,
# 	                      u_window      = sample.u_window,
# 	                      shgc_window   = sample.shgc_window,
# 	                      f_inc_hw      = sample.f_inc_hw,
# 	                      ach50         = sample.ach50,
# 	                      r_bsmtwall    = sample.bsmtwall,
# 	                      r_slab        = sample.slab))	
# }
# cat(paste("6. the ",no.sample," random samples are drawn for  CZ ",this.city,"(",this.CZ,") are stored in [df.parm]!\n",sep=""))
# cat(paste("6. the ",no.sample," random samples are drawn for  CZ ",this.city,"(",this.CZ,") are stored in [df.parm]!\n",sep=""),file=FL.LOG,append=TRUE)
# # ----------------------------------------------------------------------------------------------------------------------------------------------------------
# # August 22, 2014: the bove sampling design for detecting IECC Regulated end-use energy change is replaced by a new design to detect code item difference.
# # ----------------------------------------------------------------------------------------------------------------------------------------------------------


# # check through plotting
# id.vars       <- c("CZ_city","CZ_label","CZ_moist","CZ_weather","tag","foundation")
# measure.vars  <- parm.var.varying
# myMelt        <- melt(df.parm,id.vars=id.vars,measure.vars=measure.vars,value.name="value")
# 
# # histogram
# plot.hist <- histogram(~value | variable+CZ_city,data = myMelt,group=CZ_city,
#                        layout=c(3,3),
#                        scales=list(relation="free"))	# not sure wht free scale does not work
# plot(plot.hist)
# 
# # 
# par(mfrow = c(3,3))
# for (this.city in unique(myMelt[,"CZ_city"]))
# {
# 	for (this.var in unique(myMelt[,"variable"]))
# 	{
# 		mySub <- subset(myMelt,subset = ((variable == this.var) & (CZ_city == this.city)))
# 		hist(mySub[,"value"], main = paste(this.var," (",this.city,")",sep=""),xlab = this.var, ylab="frequence",freq=FALSE,col="grey",border="black")
# 	}
# }
# dev.off()
# cat(paste("7. histogram are plotted for each parameter at each CZ for checking aginst the input weights!!\n",sep=""))
# cat(paste("7. histogram are plotted for each parameter at each CZ for checking aginst the input weights!!\n",sep=""),file=FL.LOG,append=TRUE)
# 
# 
# # add other non-vaying parm variable and case name:
# df.parm[,"ID"] 			<- paste("SF",df.parm[,"CZ_city"],"IECC_2009",paste("MonteCarlo",row.names(df.parm),sep=""),sep="_")
# df.parm[,"weatherfile"] 	<- WZ.file[match(df.parm[,"CZ_city"],names(WZ.file))]
# df.parm[,"location"] 		<- df.parm[,"CZ_city"]
# df.parm[,"doe_climate_zone"] 	<- CZ.idx[match(df.parm[,"CZ_city"],names(CZ.idx))]
# df.parm[,"doe_moisture_regime"] <- CZ_moist[match(df.parm[,"CZ_city"],names(CZ_moist))]
# df.parm[,"prototype"] 		<- "singlefamily"
# df.parm[,"cfa_total"] 		<- 2400
# df.parm[,"code"] 		<- "IECC 2009"
# df.parm[,"heating_fuel"] 	<- "electricity"
# df.parm[grep("gasfurnace",df.parm[,"tag"]),"heating_fuel"] <- "naturalgas"
# df.parm[grep("oilfurnace",df.parm[,"tag"]),"heating_fuel"] <- "oil"
# 
# df.parm[,"code"] 		<- "IECC 2009"
# 
# # df.parm[,"n_units"]        	  <- 1
# # df.parm[,"n_stories_per_unit"]  <- 1
# # df.parm[,"heating_fuel"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# # df.parm[,"aspect_ratio"]        <- 4.71
# cat(paste("8. other parm variables are added to parm csv!!\n",sep=""))
# cat(paste("8. other parm variables are added to parm csv!!\n",sep=""),file=FL.LOG,append=TRUE)
# 
# # field output to the parm csv
# field.to.parm <- c("ID","weatherfile","location","doe_climate_zone","doe_moisture_regime","prototype","cfa_total","tag","foundation","heating_fuel","code",parm.var.varying)
# 
# # write out the parm csv file
# write.table(df.parm[,field.to.parm],file=FL.parm.2keep,sep=",",row.names=FALSE,col.names=TRUE,append=TRUE)
# cat(paste("9. the ",no.sample," random samples of all CZs are written out to ",FL.parm.2keep,"]!\n",sep=""))
# cat(paste("9. the ",no.sample," random samples of all CZs are written out to ",FL.parm.2keep,"]!\n",sep=""),file=FL.LOG,append=TRUE)
#
